Add unit tests for MITRE ATT&CK mapper and implement UI components for Streamlit application
- Created unit tests for the MITRE ATT&CK mapper agent in `test_mitre_mapper.py` to validate vulnerability mappings and structure.
- Added UI components for chat interface in `chat_interface.py`, including query processing and response handling with cancellation support.
- Developed the Pentest tab UI in `pentest_tab.py` to display scan progress and network visualization.
- Implemented sidebar settings in `sidebar.py` for model selection and Nmap settings.
- Introduced a testing tab in `testing_tab.py` for running various tests and displaying results, including message wrapper tests and performance comparisons.
- Added necessary imports and configurations to support the new UI components and testing functionalities.

YS-55
