+++ python/callbacks/logging_callbacks.py
@@ -1,0 +1,192 @@
+"""
+Custom callback handlers for detailed logging of agent operations.
+"""
+
+import logging
+import re
+import json
+import sys
+import os
+import traceback
+from langchain_core.callbacks import BaseCallbackHandler
+
+# Add the parent directory to sys.path to allow absolute imports
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+class DetailedLoggingCallbackHandler(BaseCallbackHandler):
+    """Custom callback handler for detailed logging."""
+    
+    def __init__(self):
+        super().__init__()
+        self.logger = logging.getLogger("agent_debug")
+        # Ensure the logger is set to DEBUG level
+        self.logger.setLevel(logging.DEBUG)
+        # Add a file handler for the debug logs
+        from pathlib import Path
+        logs_dir = Path(__file__).resolve().parent.parent.parent / "logs"
+        debug_log_file = logs_dir / 'agent_debug.log'
+        file_handler = logging.FileHandler(str(debug_log_file), mode='a')
+        file_handler.setFormatter(logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s'))
+        self.logger.addHandler(file_handler)
+    
+    def on_llm_start(self, serialized, prompts, **kwargs):
+        if serialized is not None and 'name' in serialized:
+            self.logger.debug(f"LLM Start: {serialized['name']}")
+        else:
+            self.logger.debug("LLM Start: unknown (serialized is None or missing 'name' key)")
+        
+        if prompts:
+            for i, prompt in enumerate(prompts):
+                if prompt:
+                    self.logger.debug(f"Prompt {i}: {prompt[:1000]}...")
+                else:
+                    self.logger.debug(f"Prompt {i}: None")
+    
+    def on_llm_end(self, response, **kwargs):
+        self.logger.debug(f"LLM Response: {response}")
+    
+    def on_llm_error(self, error, **kwargs):
+        self.logger.error(f"LLM Error: {error}")
+    
+    def on_chain_start(self, serialized, inputs, **kwargs):
+        if serialized is not None and 'name' in serialized:
+            self.logger.debug(f"Chain Start: {serialized['name']}")
+        else:
+            self.logger.debug("Chain Start: unknown (serialized is None or missing 'name' key)")
+        
+        self.logger.debug(f"Chain Inputs: {self._format_input(inputs)}")
+    
+    def on_chain_end(self, outputs, **kwargs):
+        self.logger.debug(f"Chain Outputs: {self._format_input(outputs)}")
+    
+    def on_chain_error(self, error, **kwargs):
+        self.logger.error(f"Chain Error: {error}")
+    
+    def on_tool_start(self, serialized, input_str, **kwargs):
+        if serialized is not None and 'name' in serialized:
+            self.logger.debug(f"Tool Start: {serialized['name']}")
+        else:
+            self.logger.debug("Tool Start: unknown (serialized is None or missing 'name' key)")
+        
+        self.logger.debug(f"Tool Input: {input_str}")
+        
+        # Special handling for security_analyst_tool
+        if serialized['name'] == "security_analyst":
+            self.logger.debug(f"SECURITY ANALYST TOOL CALLED WITH INPUT: {input_str}")
+            try:
+                # Try to parse the input as JSON to see its structure
+                if isinstance(input_str, str):
+                    # Check if this is an MCP tool response with truncated JSON
+                    if "toolu_" in input_str and "nmap_data" in input_str:
+                        self.logger.debug("Detected MCP tool response format, preprocessing...")
+                        # Try to extract the nmap_data JSON string
+                        nmap_data_match = re.search(r'"nmap_data":\s*"({.*?})"', input_str)
+                        if nmap_data_match:
+                            # Extract the nmap_data JSON string
+                            extracted_json = nmap_data_match.group(1)
+                            # Unescape any escaped quotes
+                            extracted_json = extracted_json.replace('\\"', '"')
+                            self.logger.debug(f"Extracted nmap_data: {extracted_json[:100]}...")
+                            
+                            # Fix truncated @services field which is a common issue in nmap output
+                            extracted_json = re.sub(r'"@services"\s*:\s*"[^"]*?(?:,\d+)?(?=\s*[,}])', r'"@services": "truncated_services"', extracted_json)
+                            
+                            # Replace truncated fields like "@services": "..." with valid JSON values
+                            extracted_json = re.sub(r'"@services"\s*:\s*"\.\.\.?"', '"@services": "truncated"', extracted_json)
+                            
+                            # Replace any other truncated fields with valid JSON values
+                            extracted_json = re.sub(r':\s*"\.\.\.?"', ': "truncated"', extracted_json)
+                            
+                            # Try to parse the preprocessed JSON
+                            try:
+                                parsed_json = json.loads(extracted_json)
+                                self.logger.debug("Successfully parsed preprocessed JSON")
+                                # Create a new input string with the preprocessed JSON
+                                new_input = json.dumps({"nmap_data": parsed_json})
+                                self.logger.debug(f"New input: {new_input[:100]}...")
+                                # Replace the original input_str with the preprocessed one
+                                input_str = new_input
+                            except json.JSONDecodeError as e:
+                                self.logger.error(f"Failed to parse preprocessed JSON: {e}")
+                                # Continue with the original input_str
+                    
+                    try:
+                        parsed = json.loads(input_str)
+                        self.logger.debug(f"Parsed input: {parsed}")
+                        if 'nmap_data' not in parsed:
+                            self.logger.error("MISSING REQUIRED PARAMETER: nmap_data not in input")
+                            self.logger.error(f"VALIDATION ERROR: security_analyst tool requires nmap_data parameter")
+                            # Log the call stack to see where this is being called from
+                            import traceback
+                            self.logger.error(f"Call stack:\n{traceback.format_stack()}")
+                    except json.JSONDecodeError:
+                        self.logger.debug(f"Input is not JSON: {input_str}")
+                        if input_str == "{}" or not input_str.strip():
+                            self.logger.error("EMPTY INPUT: security_analyst tool called with empty input")
+                else:
+                    self.logger.debug(f"Input type: {type(input_str)}")
+                    if isinstance(input_str, dict):
+                        if not input_str:
+                            self.logger.error("EMPTY DICT: security_analyst tool called with empty dictionary")
+                        elif 'nmap_data' not in input_str:
+                            self.logger.error("MISSING REQUIRED PARAMETER: nmap_data not in input dictionary")
+                            self.logger.error(f"Dictionary keys: {list(input_str.keys())}")
+                            
+                        # Check if nmap_data is a string that looks like an MCP tool response
+                        elif 'nmap_data' in input_str and isinstance(input_str['nmap_data'], str) and "toolu_" in input_str['nmap_data']:
+                            self.logger.debug("Detected MCP tool response in nmap_data, preprocessing...")
+                            # Import the preprocess_json_string function from security_analyst module
+                            from modules.security_analyst import preprocess_json_string
+                            # Preprocess the nmap_data
+                            preprocessed_data = preprocess_json_string(input_str['nmap_data'])
+                            self.logger.debug(f"Preprocessed nmap_data: {preprocessed_data[:100]}...")
+                            # Update the input_str with the preprocessed data
+                            input_str['nmap_data'] = preprocessed_data
+            except Exception as e:
+                self.logger.error(f"Error analyzing security_analyst input: {e}")
+                self.logger.error(f"Exception type: {type(e)}")
+                import traceback
+                self.logger.error(f"Exception traceback:\n{traceback.format_exc()}")
+    
+    def on_tool_end(self, output, **kwargs):
+        self.logger.debug(f"Tool Output: {output[:1000]}..." if isinstance(output, str) and len(output) > 1000 else f"Tool Output: {output}")
+    
+    def on_tool_error(self, error, **kwargs):
+        self.logger.error(f"Tool Error: {error}")
+    
+    def on_agent_action(self, action, **kwargs):
+        if action is not None and hasattr(action, 'tool'):
+            self.logger.debug(f"Agent Action: {action.tool}")
+            if hasattr(action, 'tool_input'):
+                self.logger.debug(f"Action Input: {action.tool_input}")
+            else:
+                self.logger.debug("Action Input: None (action.tool_input not available)")
+        else:
+            self.logger.debug("Agent Action: unknown (action is None or missing tool attribute)")
+        
+        # Special handling for security_analyst_tool
+        if action.tool == "security_analyst":
+            self.logger.debug(f"SECURITY ANALYST ACTION: {action.log}")
+            self.logger.debug(f"SECURITY ANALYST INPUT: {action.tool_input}")
+            
+            if not action.tool_input or (isinstance(action.tool_input, dict) and 'nmap_data' not in action.tool_input):
+                self.logger.error("AGENT TRYING TO CALL SECURITY ANALYST WITHOUT nmap_data")
+                # Log the agent's reasoning
+                self.logger.error(f"Agent reasoning: {action.log}")
+    
+    def _format_input(self, inputs):
+        """Format inputs for logging, truncating long strings"""
+        if not inputs:
+            return inputs
+        
+        try:
+            formatted = {}
+            for k, v in inputs.items():
+                if isinstance(v, str) and len(v) > 1000:
+                    formatted[k] = v[:1000] + "..."
+                else:
+                    formatted[k] = v
+            return formatted
+        except (TypeError, AttributeError):
+            # Handle case where inputs is not a dictionary or doesn't support items()
+            return str(inputs)
