+++ python/main.py
@@ -1,1000 +1,68 @@
-import json
+"""
+Main entry point for the pentest agent system.
+"""
+
 import logging
-import math
-import re
 import sys
-import time
-from concurrent.futures import ThreadPoolExecutor
+import os
 from pathlib import Path
 
 import streamlit as st
 from dotenv import load_dotenv
-from langchain.agents import AgentExecutor, create_tool_calling_agent
-from langchain_anthropic import ChatAnthropic
-from langchain_core.callbacks import BaseCallbackHandler
-from langchain_core.output_parsers import PydanticOutputParser
-from langchain_core.prompts import ChatPromptTemplate
-from langchain_openai import ChatOpenAI
-from pydantic import BaseModel
 
-# Update the global variable in tools module
-import modules.tools as tools
-from modules.metasploit_agent import metasploit_process_tool, metasploit_search_tool
-from modules.security_analyst import security_analyst_tool
-from modules.tools import nmap_tool, save_tool, search_tool, wiki_tool
+# Add the current directory to sys.path to allow absolute imports
+sys.path.insert(0, os.path.abspath(os.path.dirname(__file__)))
 
+# Import modules
+from config.settings import setup_environment
+from ui.sidebar import render_sidebar
+from ui.pentest_tab import render_pentest_tab
+from ui.testing_tab import render_testing_tab
+from ui.chat_interface import setup_chat, process_query
+from models.model_handler import get_agent_executor
+
+# Setup environment
 load_dotenv()
+logs_dir, logger = setup_environment()
 
-# Ensure logs directory exists
-logs_dir = Path(__file__).resolve().parent.parent / "logs"
-if not logs_dir.exists():
-    logs_dir.mkdir(parents=True, exist_ok=True)
-
-# Set up logging to file and console
-log_file = logs_dir / 'python_debug.log'
-logging.basicConfig(
-    level=logging.INFO,
-    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
-    handlers=[
-        logging.FileHandler(str(log_file), mode='a'),
-        logging.StreamHandler(sys.stdout)
-    ]
-)
-logger = logging.getLogger(__name__)
-
-#print("Python Orchestrator script execution started")
-logger.info("Orchestrator started")
-
-# Ensure storage directory exists
-current_dir = Path(__file__).resolve().parent
-storage_dir = current_dir.parent / "storage"
-if not storage_dir.exists():
-    logger.info(f"Creating directory: {storage_dir}")
-    storage_dir.mkdir(parents=True, exist_ok=True)
-
-# Database connection and data display code has been moved to db_debug.py
-# To test database connection, run: python db_debug.py
+# Initialize session state for scan progress if not already present
+if 'scan_progress' not in st.session_state:
+    st.session_state.scan_progress = {
+        'running': False,
+        'target': '',
+        'scan_type': '',
+        'start_time': None,
+        'percent_complete': 0,
+        'current_phase': 'Not started',
+        'ports_found': 0,
+        'services_found': 0,
+        'vulnerabilities_found': 0,
+        'status_message': '',
+        'last_update': '',
+        'estimated_time_remaining': '',
+        'log_messages': []
+    }
 
 # Create tabs for different functionality
 tab1, tab2 = st.tabs(["Pentest", "Testing"])
 
+# Render UI components
 with tab1:
-    st.title("Cybersecurity Orchestrator")
-    
-    # Check if a scan is currently running
-    if 'scan_progress' in st.session_state and st.session_state.scan_progress['running']:
-        # Create a container for the scan progress UI
-        scan_progress_container = st.container()
-        
-        with scan_progress_container:
-            st.subheader(f"Scanning {st.session_state.scan_progress['target']} ({st.session_state.scan_progress['scan_type']} scan)")
-            
-            # Progress bar
-            st.progress(st.session_state.scan_progress['percent_complete'] / 100)
-            
-            # Status information
-            col1, col2, col3 = st.columns(3)
-            with col1:
-                st.metric("Current Phase", st.session_state.scan_progress['current_phase'])
-            with col2:
-                st.metric("Open Ports", st.session_state.scan_progress['ports_found'])
-            with col3:
-                st.metric("Vulnerabilities", st.session_state.scan_progress['vulnerabilities_found'])
-            
-            # Time information
-            if st.session_state.scan_progress['start_time']:
-                elapsed = time.time() - st.session_state.scan_progress['start_time']
-                elapsed_str = f"{int(elapsed // 60)}m {int(elapsed % 60)}s"
-                st.info(f"⏱️ Elapsed: {elapsed_str} | ⏳ {st.session_state.scan_progress['estimated_time_remaining']}")
-            
-            # Latest update
-            if st.session_state.scan_progress['last_update']:
-                st.success(f"Latest: {st.session_state.scan_progress['last_update']}")
-            
-            # Scan log
-            with st.expander("Scan Log", expanded=False):
-                for log_entry in st.session_state.scan_progress['log_messages']:
-                    st.text(log_entry)
-            
-            # Scan type information
-            scan_type = st.session_state.scan_progress['scan_type']
-            if scan_type in tools.SCAN_TYPE_INFO:
-                with st.expander("About This Scan Type", expanded=False):
-                    info = tools.SCAN_TYPE_INFO[scan_type]
-                    st.markdown(f"### {scan_type.capitalize()} Scan")
-                    st.markdown(f"**Description:** {info['description']}")
-                    st.markdown(f"**Details:** {info['details']}")
-                    st.markdown(f"**Typical Duration:** {info['duration']}")
-                    st.markdown(f"**Detection Risk:** {info['detection_risk']}")
-            
-            # Cybersecurity tips
-            with st.expander("Cybersecurity Tips While You Wait", expanded=True):
-                # Display a random tip that changes every few seconds
-                tip_index = int((time.time() // 10) % len(tools.CYBERSECURITY_TIPS))
-                st.info(f"💡 **Tip:** {tools.CYBERSECURITY_TIPS[tip_index]}")
-            
-            # Network visualization
-            with st.expander("Network Visualization", expanded=True):
-                # Create a simple network visualization using HTML/CSS
-                ports = st.session_state.scan_progress['ports_found']
-                vulnerabilities = st.session_state.scan_progress['vulnerabilities_found']
-                
-                # Generate HTML for the visualization
-                html = f"""
-                <style>
-                    .network-viz {{
-                        display: flex;
-                        justify-content: center;
-                        align-items: center;
-                        height: 300px;
-                        position: relative;
-                    }}
-                    .target {{
-                        width: 100px;
-                        height: 100px;
-                        background-color: #1E88E5;
-                        border-radius: 50%;
-                        display: flex;
-                        justify-content: center;
-                        align-items: center;
-                        color: white;
-                        font-weight: bold;
-                        z-index: 10;
-                        box-shadow: 0 0 20px rgba(30, 136, 229, 0.5);
-                    }}
-                    .port {{
-                        position: absolute;
-                        width: 30px;
-                        height: 30px;
-                        background-color: #4CAF50;
-                        border-radius: 50%;
-                        display: flex;
-                        justify-content: center;
-                        align-items: center;
-                        color: white;
-                        font-size: 10px;
-                        box-shadow: 0 0 10px rgba(76, 175, 80, 0.5);
-                    }}
-                    .vuln {{
-                        position: absolute;
-                        width: 40px;
-                        height: 40px;
-                        background-color: #F44336;
-                        border-radius: 50%;
-                        display: flex;
-                        justify-content: center;
-                        align-items: center;
-                        color: white;
-                        font-size: 12px;
-                        box-shadow: 0 0 15px rgba(244, 67, 54, 0.7);
-                        animation: pulse 2s infinite;
-                    }}
-                    @keyframes pulse {{
-                        0% {{ transform: scale(1); }}
-                        50% {{ transform: scale(1.1); }}
-                        100% {{ transform: scale(1); }}
-                    }}
-                    .line {{
-                        position: absolute;
-                        background-color: #90A4AE;
-                        height: 2px;
-                        transform-origin: 0 0;
-                        z-index: 1;
-                    }}
-                </style>
-                <div class="network-viz">
-                    <div class="target">{st.session_state.scan_progress['target']}</div>
-                """
-                
-                # Add ports (green circles)
-                import random
-                random.seed(42)  # Use fixed seed for consistent layout
-                for i in range(min(ports, 10)):  # Show up to 10 ports
-                    angle = i * (360 / min(ports, 10)) * (3.14159 / 180)  # Convert to radians
-                    distance = 120  # Distance from center
-                    x = 50 + distance * math.cos(angle)
-                    y = 50 + distance * math.sin(angle)
-                    
-                    # Calculate line properties
-                    length = math.sqrt((x - 50) ** 2 + (y - 50) ** 2)
-                    angle_deg = math.atan2(y - 50, x - 50) * (180 / 3.14159)
-                    
-                    # Add line
-                    html += f"""
-                    <div class="line" style="
-                        width: {length}px;
-                        left: 50%;
-                        top: 50%;
-                        transform: rotate({angle_deg}deg);
-                    "></div>
-                    """
-                    
-                    # Add port
-                    html += f"""
-                    <div class="port" style="left: calc(50% + {x - 15}px); top: calc(50% + {y - 15}px);">
-                        {i+1}
-                    </div>
-                    """
-                
-                # Add vulnerabilities (red circles)
-                for i in range(min(vulnerabilities, 5)):  # Show up to 5 vulnerabilities
-                    angle = i * (360 / min(vulnerabilities, 5)) * (3.14159 / 180)  # Convert to radians
-                    distance = 180  # Distance from center
-                    x = 50 + distance * math.cos(angle)
-                    y = 50 + distance * math.sin(angle)
-                    
-                    # Calculate line properties
-                    length = math.sqrt((x - 50) ** 2 + (y - 50) ** 2)
-                    angle_deg = math.atan2(y - 50, x - 50) * (180 / 3.14159)
-                    
-                    # Add line
-                    html += f"""
-                    <div class="line" style="
-                        width: {length}px;
-                        left: 50%;
-                        top: 50%;
-                        transform: rotate({angle_deg}deg);
-                    "></div>
-                    """
-                    
-                    # Add vulnerability
-                    html += f"""
-                    <div class="vuln" style="left: calc(50% + {x - 20}px); top: calc(50% + {y - 20}px);">
-                        !
-                    </div>
-                    """
-                
-                html += "</div>"
-                
-                # Display the visualization
-                st.components.v1.html(html, height=300)
-            
-            # Auto-refresh the UI every few seconds
-            st.empty()
-            time.sleep(2)
-            st.experimental_rerun()
+    render_pentest_tab()
 
 with tab2:
-    st.title("Security Analyst Testing")
-    
-    # Import test modules
-    from tests.mock_search import mock_search
-    from modules.security_analyst import analyze_nmap_results
-    
-    # Initialize session state for test results
-    if "test_results" not in st.session_state:
-        st.session_state.test_results = None
-    if "test_running" not in st.session_state:
-        st.session_state.test_running = False
-    
-    # Test configuration
-    st.subheader("Test Configuration")
-    
-    col1, col2 = st.columns(2)
-    
-    with col1:
-        test_type = st.selectbox(
-            "Select Test Type",
-            ["Direct Test", "Mock Search Test", "Real Search Test", "Compare Tests", "Message Wrapper Test", "Real IP Test"],
-            index=0,
-            key="test_type"
-        )
-        
-        sample_file = st.selectbox(
-            "Select Sample Data",
-            ["sample_nmap_output.txt", "Upload Custom File"],
-            index=0,
-            key="sample_file"
-        )
-        
-        if sample_file == "Upload Custom File":
-            uploaded_file = st.file_uploader("Upload Nmap Output File", type=["txt", "xml", "json"])
-        
-    with col2:
-        st.markdown("### Test Options")
-        use_mock = st.checkbox("Use Mock Search", value=True, key="use_mock")
-        show_logs = st.checkbox("Show Debug Logs", value=False, key="show_logs")
-        compare_performance = st.checkbox("Compare Performance", value=False, key="compare_performance")
-    
-    # Message Wrapper Test Configuration UI (shown conditionally)
-    if test_type == "Message Wrapper Test":
-        from modules.security_analyst import preprocess_json_string
-        
-        st.markdown("### Message Wrapper Test Configuration")
-        st.info("This test checks the system's ability to handle LangChain message wrappers like AIMessageChunk.")
-        
-        wrapper_type = st.selectbox(
-            "Select Wrapper Type",
-            ["AIMessageChunk", "HumanMessage", "SystemMessage", "Custom"],
-            index=0,
-            key="wrapper_type"
-        )
-        
-        # Sample JSON content to wrap
-        default_json = """{
-  "format": "json",
-  "scan_type": "vuln",
-  "target": "192.168.1.100",
-  "text_output": "NMAP SCAN RESULTS (vuln scan of 192.168.1.100):\\n==================================================\\n...\\n| smb-vuln-ms17-010: \\n|   VULNERABLE:\\n|   Remote Code Execution vulnerability in Microsoft SMBv1 servers (ms17-010)\\n|     State: VULNERABLE"
-}"""
-        
-        json_content = st.text_area(
-            "JSON Content to Wrap",
-            value=default_json,
-            height=200,
-            key="json_content"
-        )
-        
-        # Create the wrapped content based on selected wrapper type
-        if wrapper_type == "Custom":
-            custom_wrapper = st.text_input(
-                "Custom Wrapper Format (use {} as placeholder for content)",
-                value="CustomMessage(content={})",
-                key="custom_wrapper"
-            )
-            wrapped_content = custom_wrapper.format(f'"{json_content}"')
-        else:
-            wrapped_content = f'{wrapper_type}(content="{json_content}")'
-        
-        # Show the wrapped content
-        with st.expander("View Wrapped Content", expanded=False):
-            st.code(wrapped_content, language="python")
-    
-    # Run test button
-    if st.button("Run Test", type="primary", key="run_test"):
-        st.session_state.test_running = True
-        
-        with st.spinner("Running test..."):
-            try:
-                # For non-Message Wrapper tests, load sample data
-                if test_type != "Message Wrapper Test":
-                    if sample_file == "Upload Custom File" and uploaded_file is not None:
-                        sample_data = uploaded_file.read().decode()
-                    else:
-                        sample_file_path = Path(__file__).resolve().parent / "tests" / sample_file
-                        with open(sample_file_path, "r") as f:
-                            sample_data = f.read()
-                
-                # Process based on test type
-                if test_type == "Message Wrapper Test":
-                    # Process the wrapped content that was configured above
-                    start_time = time.time()
-                    preprocessed = preprocess_json_string(wrapped_content)
-                    
-                    # Try to parse the preprocessed JSON
-                    try:
-                        parsed_json = json.loads(preprocessed)
-                        
-                        # Store results
-                        result = {
-                            "success": True,
-                            "preprocessing_time": time.time() - start_time,
-                            "parsed_json": parsed_json,
-                            "wrapper_type": wrapper_type,
-                            "original_wrapped": wrapped_content,
-                            "preprocessed": preprocessed,
-                            "is_message_wrapper_test": True
-                        }
-                    except json.JSONDecodeError as e:
-                        result = {
-                            "success": False,
-                            "error": f"JSON parsing error: {str(e)}",
-                            "error_position": e.pos,
-                            "error_context": preprocessed[max(0, e.pos-50):min(len(preprocessed), e.pos+50)],
-                            "wrapper_type": wrapper_type,
-                            "original_wrapped": wrapped_content,
-                            "preprocessed": preprocessed,
-                            "is_message_wrapper_test": True
-                        }
-                
-                elif test_type == "Real IP Test":
-                    # Load the sample data for 10.0.255.165 from the test_real_ip.py script
-                    real_ip_sample_path = Path(__file__).resolve().parent / "test_real_ip.py"
-                    
-                    # Extract the sample data from the test_real_ip.py file
-                    with open(real_ip_sample_path, "r") as f:
-                        file_content = f.read()
-                        # Extract the sample data between triple quotes
-                        import re
-                        sample_data_match = re.search(r'sample_data = """(.*?)"""', file_content, re.DOTALL)
-                        if sample_data_match:
-                            sample_data = sample_data_match.group(1)
-                            st.info("Testing with real IP: 10.0.255.165")
-                        else:
-                            st.error("Could not extract sample data from test_real_ip.py")
-                            sample_data = "Error: Could not extract sample data"
-                    
-                    # Set start time for duration calculation
-                    start_time = time.time()
-                    
-                    # Analyze the sample data
-                    if use_mock:
-                        result = analyze_nmap_results(sample_data, custom_search_func=mock_search)
-                    else:
-                        result = analyze_nmap_results(sample_data)
-                
-                elif test_type == "Direct Test":
-                    if use_mock:
-                        result = analyze_nmap_results(sample_data, custom_search_func=mock_search)
-                    else:
-                        result = analyze_nmap_results(sample_data)
-                    
-                elif test_type == "Compare Tests":
-                    # Run both real and mock tests and compare
-                    mock_start_time = time.time()
-                    mock_result = analyze_nmap_results(sample_data, custom_search_func=mock_search)
-                    mock_duration = time.time() - mock_start_time
-                    
-                    real_start_time = time.time()
-                    real_result = analyze_nmap_results(sample_data)
-                    real_duration = time.time() - real_start_time
-                    
-                    result = {
-                        "mock_result": mock_result,
-                        "real_result": real_result,
-                        "mock_duration": mock_duration,
-                        "real_duration": real_duration,
-                        "is_comparison": True
-                    }
-                else:
-                    # Default to direct test
-                    if use_mock:
-                        result = analyze_nmap_results(sample_data, custom_search_func=mock_search)
-                    else:
-                        result = analyze_nmap_results(sample_data)
-                
-                duration = time.time() - start_time
-                
-                # Store results in session state
-                st.session_state.test_results = {
-                    "result": result,
-                    "duration": duration,
-                    "test_type": test_type,
-                    "use_mock": use_mock,
-                    "sample_file": sample_file
-                }
-                
-            except Exception as e:
-                st.error(f"Error running test: {str(e)}")
-                st.session_state.test_results = {
-                    "error": str(e),
-                    "test_type": test_type,
-                    "use_mock": use_mock,
-                    "sample_file": sample_file
-                }
-            
-            st.session_state.test_running = False
-    
-    # Display test results
-    if st.session_state.test_results:
-        st.subheader("Test Results")
-        
-        results = st.session_state.test_results
-        
-        if "error" in results:
-            st.error(f"Test failed: {results['error']}")
-        else:
-            # Display test metadata
-            st.info(f"Test completed in {results['duration']:.2f} seconds")
-            
-            if "is_message_wrapper_test" in results["result"] and results["result"]["is_message_wrapper_test"]:
-                # Display Message Wrapper Test results
-                wrapper_result = results["result"]
-                
-                st.markdown("### Message Wrapper Test Results")
-                
-                if wrapper_result.get("success", False):
-                    st.success(f"✅ Successfully processed {wrapper_result['wrapper_type']} wrapper in {wrapper_result['preprocessing_time']:.4f} seconds")
-                    
-                    # Display the parsed JSON
-                    st.subheader("Parsed JSON")
-                    st.json(wrapper_result["parsed_json"])
-                    
-                    # Show before and after
-                    col1, col2 = st.columns(2)
-                    
-                    with col1:
-                        st.markdown("#### Original Wrapped Content")
-                        st.code(wrapper_result["original_wrapped"], language="python")
-                    
-                    with col2:
-                        st.markdown("#### Preprocessed Content")
-                        st.code(wrapper_result["preprocessed"], language="json")
-                    
-                    # Verify key fields were preserved
-                    if "parsed_json" in wrapper_result:
-                        st.subheader("Field Verification")
-                        parsed = wrapper_result["parsed_json"]
-                        
-                        # Check for common fields
-                        fields_to_check = ["format", "target", "scan_type"]
-                        for field in fields_to_check:
-                            if field in parsed:
-                                st.success(f"✅ Field '{field}' was preserved: {parsed[field]}")
-                            else:
-                                st.error(f"❌ Field '{field}' was not preserved")
-                else:
-                    st.error(f"❌ Failed to process {wrapper_result['wrapper_type']} wrapper")
-                    st.error(wrapper_result.get("error", "Unknown error"))
-                    
-                    if "error_context" in wrapper_result:
-                        st.markdown("#### Error Context")
-                        st.code(f"Error at position {wrapper_result['error_position']}:\n{wrapper_result['error_context']}")
-                    
-                    if "preprocessed" in wrapper_result:
-                        st.markdown("#### Preprocessed Content (Invalid JSON)")
-                        st.code(wrapper_result["preprocessed"])
-                    
-                    # Show the original wrapped content
-                    st.markdown("#### Original Wrapped Content")
-                    st.code(wrapper_result["original_wrapped"], language="python")
-            
-            elif "is_comparison" in results["result"] and results["result"]["is_comparison"]:
-                # Display comparison results
-                comparison = results["result"]
-                
-                st.markdown("### Performance Comparison")
-                col1, col2 = st.columns(2)
-                
-                with col1:
-                    st.markdown("#### Mock Search")
-                    st.info(f"Duration: {comparison['mock_duration']:.2f} seconds")
-                
-                with col2:
-                    st.markdown("#### Real Search")
-                    st.info(f"Duration: {comparison['real_duration']:.2f} seconds")
-                
-                st.markdown("### MITRE ATT&CK Mappings Comparison")
-                
-                col1, col2 = st.columns(2)
-                
-                with col1:
-                    st.markdown("#### Mock Search Mappings")
-                    if "mitre_mappings" in comparison["mock_result"]:
-                        for i, mapping in enumerate(comparison["mock_result"]["mitre_mappings"]):
-                            st.markdown(f"**{i+1}. Technique:** {mapping.get('technique_id')} - {mapping.get('technique_name')}")
-                            st.markdown(f"**Tactic:** {mapping.get('tactic')}")
-                            st.markdown("---")
-                
-                with col2:
-                    st.markdown("#### Real Search Mappings")
-                    if "mitre_mappings" in comparison["real_result"]:
-                        for i, mapping in enumerate(comparison["real_result"]["mitre_mappings"]):
-                            st.markdown(f"**{i+1}. Technique:** {mapping.get('technique_id')} - {mapping.get('technique_name')}")
-                            st.markdown(f"**Tactic:** {mapping.get('tactic')}")
-                            st.markdown("---")
-                
-                # Display vulnerabilities
-                st.markdown("### Vulnerabilities")
-                if "vulnerabilities" in comparison["mock_result"]:
-                    with st.expander("View Vulnerabilities", expanded=True):
-                        st.json(comparison["mock_result"]["vulnerabilities"])
-                
-                # Display attack plan
-                st.markdown("### Attack Plan")
-                if "attack_plan" in comparison["mock_result"]:
-                    with st.expander("View Attack Plan", expanded=True):
-                        # Display executive summary
-                        if "executive_summary" in comparison["mock_result"]["attack_plan"]:
-                            st.subheader("Executive Summary")
-                            st.write(comparison["mock_result"]["attack_plan"]["executive_summary"])
-                        
-                        # Display task list
-                        if "task_list" in comparison["mock_result"]["attack_plan"]:
-                            st.subheader("Task List")
-                            for task in comparison["mock_result"]["attack_plan"]["task_list"]:
-                                st.markdown(f"**Task {task.get('id', '')}:** {task.get('name', '')}")
-                                st.markdown(f"*Description:* {task.get('description', '')}")
-                                if "commands" in task:
-                                    st.code("\n".join(task["commands"]), language="bash")
-                                st.markdown("---")
-            else:
-                # Display regular test results
-                result = results["result"]
-                
-                # Display vulnerabilities
-                if "vulnerabilities" in result:
-                    st.markdown("### Vulnerabilities Found")
-                    with st.expander("View Vulnerabilities", expanded=True):
-                        for i, vuln in enumerate(result["vulnerabilities"]):
-                            st.markdown(f"**{i+1}. {vuln.get('name')}**")
-                            st.markdown(f"*Description:* {vuln.get('description')}")
-                            st.markdown(f"*ID:* {vuln.get('id')}")
-                            st.markdown("---")
-                
-                # Display MITRE mappings
-                if "mitre_mappings" in result:
-                    st.markdown("### MITRE ATT&CK Mappings")
-                    with st.expander("View MITRE Mappings", expanded=True):
-                        for i, mapping in enumerate(result["mitre_mappings"]):
-                            st.markdown(f"**{i+1}. Technique:** {mapping.get('technique_id')} - {mapping.get('technique_name')}")
-                            st.markdown(f"**Tactic:** {mapping.get('tactic')}")
-                            st.markdown("---")
-                
-                # Display attack plan
-                if "attack_plan" in result:
-                    st.markdown("### Attack Plan")
-                    with st.expander("View Attack Plan", expanded=True):
-                        # Display executive summary
-                        if "executive_summary" in result["attack_plan"]:
-                            st.subheader("Executive Summary")
-                            st.write(result["attack_plan"]["executive_summary"])
-                        
-                        # Display detailed plan
-                        if "detailed_plan" in result["attack_plan"]:
-                            st.subheader("Detailed Plan")
-                            st.markdown(result["attack_plan"]["detailed_plan"])
-                        
-                        # Display task list
-                        if "task_list" in result["attack_plan"]:
-                            st.subheader("Task List")
-                            for task in result["attack_plan"]["task_list"]:
-                                st.markdown(f"**Task {task.get('id', '')}:** {task.get('name', '')}")
-                                st.markdown(f"*Description:* {task.get('description', '')}")
-                                if "commands" in task:
-                                    st.code("\n".join(task["commands"]), language="bash")
-                                st.markdown("---")
-                
-                # Display raw scan output
-                if "raw_scan_output" in result:
-                    st.markdown("### Raw Scan Output")
-                    with st.expander("View Raw Scan Output", expanded=False):
-                        st.code(result["raw_scan_output"], language="bash")
-                
-                # Display raw result for debugging
-                if show_logs:
-                    st.markdown("### Debug Information")
-                    with st.expander("View Raw Result Object", expanded=False):
-                        st.json(result)
-
-# Initialize session state for chat history
-if "messages" not in st.session_state:
-    st.session_state.messages = []
+    render_testing_tab()
 
 # Initialize session state for model selection
 if "model_option" not in st.session_state:
     st.session_state.model_option = "Claude (Anthropic)"
 
-# Initialize session state variables for cancellation
-if "is_processing" not in st.session_state:
-    st.session_state.is_processing = False
-if "cancel_requested" not in st.session_state:
-    st.session_state.cancel_requested = False
-
-# Add model selection in sidebar
-st.sidebar.title("Settings")
-model_option = st.sidebar.selectbox(
-    "Select AI Model",
-    ["Claude (Anthropic)", "GPT-4o mini (OpenAI)"],
-    index=0,
-    key="model_selector"
-)
-
-# Add nmap settings in sidebar
-st.sidebar.divider()
-st.sidebar.subheader("Nmap Settings")
-use_docker = st.sidebar.checkbox("Use Docker for Nmap", value=False,
-                               help="Run nmap in a Docker container instead of locally")
-st.sidebar.info("All scans use non-privileged techniques that don't require root access")
-
-tools.USE_DOCKER_NMAP = use_docker
+# Setup sidebar and get model option
+model_option = render_sidebar()
 
 # Track if model has changed
 model_changed = st.session_state.model_option != model_option
 st.session_state.model_option = model_option
-
-class ModelResponse(BaseModel):
-    topic: str
-    summary: str
-    sources: list[str]
-    tools_used: list[str]
-    scan_results: str = ""  # Optional field for nmap results
-    vulnerabilities: list = []  # List of vulnerabilities found
-    attack_plan: dict = {}  # Attack plan generated by security analyst
-    exploit_paths: list = []  # List of exploit paths found by metasploit agent
-    
-    @classmethod
-    def parse_obj(cls, obj):
-        """Custom parser to handle various input formats"""
-        try:
-            # If obj is a list, try to parse the first item
-            if isinstance(obj, list):
-                if obj and isinstance(obj[0], dict):
-                    return super().parse_obj(obj[0])
-                # If the list is empty or doesn't contain a dict, create a minimal valid response
-                return cls(
-                    topic="Error parsing response",
-                    summary="The response could not be parsed correctly.",
-                    sources=["N/A"],
-                    tools_used=["N/A"]
-                )
-            # If obj is a string, try to parse it as JSON
-            elif isinstance(obj, str):
-                try:
-                    parsed_obj = json.loads(obj)
-                    return cls.parse_obj(parsed_obj)
-                except json.JSONDecodeError:
-                    # If it's not valid JSON, create a minimal valid response
-                    return cls(
-                        topic="Error parsing response",
-                        summary=f"Could not parse response as JSON: {obj[:100]}...",
-                        sources=["N/A"],
-                        tools_used=["N/A"]
-                    )
-            # Otherwise, use the default parser
-            return super().parse_obj(obj)
-        except Exception as e:
-            # Fallback for any other parsing errors
-            logger.error(f"Error parsing ModelResponse: {e}")
-            return cls(
-                topic="Error parsing response",
-                summary=f"An error occurred while parsing the response: {str(e)}",
-                sources=["N/A"],
-                tools_used=["N/A"]
-            )
-
-# Custom callback handler for detailed logging
-class DetailedLoggingCallbackHandler(BaseCallbackHandler):
-    def __init__(self):
-        super().__init__()
-        self.logger = logging.getLogger("agent_debug")
-        # Ensure the logger is set to DEBUG level
-        self.logger.setLevel(logging.DEBUG)
-        # Add a file handler for the debug logs
-        debug_log_file = logs_dir / 'agent_debug.log'
-        file_handler = logging.FileHandler(str(debug_log_file), mode='a')
-        file_handler.setFormatter(logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s'))
-        self.logger.addHandler(file_handler)
-    
-    def on_llm_start(self, serialized, prompts, **kwargs):
-        if serialized is not None and 'name' in serialized:
-            self.logger.debug(f"LLM Start: {serialized['name']}")
-        else:
-            self.logger.debug("LLM Start: unknown (serialized is None or missing 'name' key)")
-        
-        if prompts:
-            for i, prompt in enumerate(prompts):
-                if prompt:
-                    self.logger.debug(f"Prompt {i}: {prompt[:1000]}...")
-                else:
-                    self.logger.debug(f"Prompt {i}: None")
-    
-    def on_llm_end(self, response, **kwargs):
-        self.logger.debug(f"LLM Response: {response}")
-    
-    def on_llm_error(self, error, **kwargs):
-        self.logger.error(f"LLM Error: {error}")
-    
-    def on_chain_start(self, serialized, inputs, **kwargs):
-        if serialized is not None and 'name' in serialized:
-            self.logger.debug(f"Chain Start: {serialized['name']}")
-        else:
-            self.logger.debug("Chain Start: unknown (serialized is None or missing 'name' key)")
-        
-        self.logger.debug(f"Chain Inputs: {self._format_input(inputs)}")
-    
-    def on_chain_end(self, outputs, **kwargs):
-        self.logger.debug(f"Chain Outputs: {self._format_input(outputs)}")
-    
-    def on_chain_error(self, error, **kwargs):
-        self.logger.error(f"Chain Error: {error}")
-    
-    def on_tool_start(self, serialized, input_str, **kwargs):
-        if serialized is not None and 'name' in serialized:
-            self.logger.debug(f"Tool Start: {serialized['name']}")
-        else:
-            self.logger.debug("Tool Start: unknown (serialized is None or missing 'name' key)")
-        
-        self.logger.debug(f"Tool Input: {input_str}")
-        
-        # Special handling for security_analyst_tool
-        if serialized['name'] == "security_analyst":
-            self.logger.debug(f"SECURITY ANALYST TOOL CALLED WITH INPUT: {input_str}")
-            try:
-                # Try to parse the input as JSON to see its structure
-                if isinstance(input_str, str):
-                    # Check if this is an MCP tool response with truncated JSON
-                    if "toolu_" in input_str and "nmap_data" in input_str:
-                        self.logger.debug("Detected MCP tool response format, preprocessing...")
-                        # Try to extract the nmap_data JSON string
-                        nmap_data_match = re.search(r'"nmap_data":\s*"({.*?})"', input_str)
-                        if nmap_data_match:
-                            # Extract the nmap_data JSON string
-                            extracted_json = nmap_data_match.group(1)
-                            # Unescape any escaped quotes
-                            extracted_json = extracted_json.replace('\\"', '"')
-                            self.logger.debug(f"Extracted nmap_data: {extracted_json[:100]}...")
-                            
-                            # Fix truncated @services field which is a common issue in nmap output
-                            extracted_json = re.sub(r'"@services"\s*:\s*"[^"]*?(?:,\d+)?(?=\s*[,}])', r'"@services": "truncated_services"', extracted_json)
-                            
-                            # Replace truncated fields like "@services": "..." with valid JSON values
-                            extracted_json = re.sub(r'"@services"\s*:\s*"\.\.\.?"', '"@services": "truncated"', extracted_json)
-                            
-                            # Replace any other truncated fields with valid JSON values
-                            extracted_json = re.sub(r':\s*"\.\.\.?"', ': "truncated"', extracted_json)
-                            
-                            # Try to parse the preprocessed JSON
-                            try:
-                                parsed_json = json.loads(extracted_json)
-                                self.logger.debug("Successfully parsed preprocessed JSON")
-                                # Create a new input string with the preprocessed JSON
-                                new_input = json.dumps({"nmap_data": parsed_json})
-                                self.logger.debug(f"New input: {new_input[:100]}...")
-                                # Replace the original input_str with the preprocessed one
-                                input_str = new_input
-                            except json.JSONDecodeError as e:
-                                self.logger.error(f"Failed to parse preprocessed JSON: {e}")
-                                # Continue with the original input_str
-                    
-                    try:
-                        parsed = json.loads(input_str)
-                        self.logger.debug(f"Parsed input: {parsed}")
-                        if 'nmap_data' not in parsed:
-                            self.logger.error("MISSING REQUIRED PARAMETER: nmap_data not in input")
-                            self.logger.error(f"VALIDATION ERROR: security_analyst tool requires nmap_data parameter")
-                            # Log the call stack to see where this is being called from
-                            import traceback
-                            self.logger.error(f"Call stack:\n{traceback.format_stack()}")
-                    except json.JSONDecodeError:
-                        self.logger.debug(f"Input is not JSON: {input_str}")
-                        if input_str == "{}" or not input_str.strip():
-                            self.logger.error("EMPTY INPUT: security_analyst tool called with empty input")
-                else:
-                    self.logger.debug(f"Input type: {type(input_str)}")
-                    if isinstance(input_str, dict):
-                        if not input_str:
-                            self.logger.error("EMPTY DICT: security_analyst tool called with empty dictionary")
-                        elif 'nmap_data' not in input_str:
-                            self.logger.error("MISSING REQUIRED PARAMETER: nmap_data not in input dictionary")
-                            self.logger.error(f"Dictionary keys: {list(input_str.keys())}")
-                            
-                        # Check if nmap_data is a string that looks like an MCP tool response
-                        elif 'nmap_data' in input_str and isinstance(input_str['nmap_data'], str) and "toolu_" in input_str['nmap_data']:
-                            self.logger.debug("Detected MCP tool response in nmap_data, preprocessing...")
-                            # Import the preprocess_json_string function from security_analyst module
-                            from modules.security_analyst import preprocess_json_string
-                            # Preprocess the nmap_data
-                            preprocessed_data = preprocess_json_string(input_str['nmap_data'])
-                            self.logger.debug(f"Preprocessed nmap_data: {preprocessed_data[:100]}...")
-                            # Update the input_str with the preprocessed data
-                            input_str['nmap_data'] = preprocessed_data
-            except Exception as e:
-                self.logger.error(f"Error analyzing security_analyst input: {e}")
-                self.logger.error(f"Exception type: {type(e)}")
-                import traceback
-                self.logger.error(f"Exception traceback:\n{traceback.format_exc()}")
-    
-    def on_tool_end(self, output, **kwargs):
-        self.logger.debug(f"Tool Output: {output[:1000]}..." if isinstance(output, str) and len(output) > 1000 else f"Tool Output: {output}")
-    
-    def on_tool_error(self, error, **kwargs):
-        self.logger.error(f"Tool Error: {error}")
-    
-    def on_agent_action(self, action, **kwargs):
-        if action is not None and hasattr(action, 'tool'):
-            self.logger.debug(f"Agent Action: {action.tool}")
-            if hasattr(action, 'tool_input'):
-                self.logger.debug(f"Action Input: {action.tool_input}")
-            else:
-                self.logger.debug("Action Input: None (action.tool_input not available)")
-        else:
-            self.logger.debug("Agent Action: unknown (action is None or missing tool attribute)")
-        
-        # Special handling for security_analyst_tool
-        if action.tool == "security_analyst":
-            self.logger.debug(f"SECURITY ANALYST ACTION: {action.log}")
-            self.logger.debug(f"SECURITY ANALYST INPUT: {action.tool_input}")
-            
-            if not action.tool_input or (isinstance(action.tool_input, dict) and 'nmap_data' not in action.tool_input):
-                self.logger.error("AGENT TRYING TO CALL SECURITY ANALYST WITHOUT nmap_data")
-                # Log the agent's reasoning
-                self.logger.error(f"Agent reasoning: {action.log}")
-    
-    def _format_input(self, inputs):
-        """Format inputs for logging, truncating long strings"""
-        if not inputs:
-            return inputs
-        
-        try:
-            formatted = {}
-            for k, v in inputs.items():
-                if isinstance(v, str) and len(v) > 1000:
-                    formatted[k] = v[:1000] + "..."
-                else:
-                    formatted[k] = v
-            return formatted
-        except (TypeError, AttributeError):
-            # Handle case where inputs is not a dictionary or doesn't support items()
-            return str(inputs)
-
-# Function to create agent with selected model
-@st.cache_resource(show_spinner=False)
-def get_agent_executor(model_name):
-    if model_name == "Claude (Anthropic)":
-        llm = ChatAnthropic(model="claude-3-5-sonnet-20241022")
-    else:
-        llm = ChatOpenAI(temperature=0, model="gpt-4o-mini")
-
-    parser = PydanticOutputParser(pydantic_object=ModelResponse)
-    
-    # Create the callback handler
-    debug_callbacks = [DetailedLoggingCallbackHandler()]
-
-    prompt = ChatPromptTemplate.from_messages(
-        [
-            (
-                "system",
-                """
-                You are a senior cybersecurity research analyst who will help pentest a target system.
-                You have access to several tools:
-
-                1. scan_target (MCP) - Scan a target IP address or hostname with nmap. You can specify scan types:
-                   - basic: Standard TCP scan with version detection (default)
-                   - quick: Fast TCP scan of common ports
-                   - full: Comprehensive TCP scan of all ports with version detection
-                   - vuln: TCP scan with vulnerability scripts
-                   - udp_ports: TCP scan of common UDP service ports
-                   - version: Get nmap version information
-
-                   Example usage: scan_target(target="192.168.1.1", scan_type="vuln")
-
-                2. analyze_scan (MCP) - Analyze the most recent scan results to identify vulnerabilities, map them to MITRE ATT&CK framework, and generate an attack plan.
-                   IMPORTANT: This tool must only be used after running scan_target.
-                   Example usage: analyze_scan()
-
-                3. metasploit_search - Search for Metasploit exploits related to a vulnerability. Input can be a vulnerability name or a JSON object with vulnerability details.
-
-                4. metasploit_process_latest - Process the latest vulnerabilities from the database and search for Metasploit exploits for each one.
-
-                5. search - Search the web for cybersecurity information
-
-                6. wiki_tool - Query Wikipedia for relevant information
-
-                7. save_text_to_file - Save your findings to a file
-
-                When scanning targets, you MUST ALWAYS include the COMPLETE and UNMODIFIED nmap scan results in your response.
-                Never summarize or omit any part of the nmap scan output - the full raw output must be preserved.
-
-                Be thorough in your analysis of scan results, identifying potential vulnerabilities.
-
-                CRITICAL INSTRUCTION FOR SECURITY_ANALYST TOOL:
-                When a vulnerability is found (indicated by "State: VULNERABLE" in nmap results), use the security_analyst tool to analyze it, but you MUST ALWAYS provide the nmap_data parameter with the scan results. 
-                NEVER call security_analyst with an empty object {{}} or without the nmap_data parameter.
-                
-                CORRECT USAGE: security_analyst(nmap_data=scan_results)
-                INCORRECT USAGE: security_analyst() or security_analyst({{}})
-                
-                After analysis, use the metasploit_search tool to find relevant exploits.
-
-                IMPORTANT: Always set the scan_results field in your response to the complete nmap output.
-                If you use the security_analyst tool, use the raw_scan_output field from its response as your scan_results.
-
-                Previous conversation: {chat_history}
-
-                Answer the user query and use necessary tools.
-                Wrap the output in this format and provide no other text\n
-                {format_instructions}
-                """,
-            ),
-            ("human", "{query}"),
-            ("ai", "{agent_scratchpad}"),
-        ]
-    ).partial(format_instructions=parser.get_format_instructions())
-
-    tools = [search_tool, wiki_tool, save_tool, nmap_tool, security_analyst_tool, metasploit_search_tool, metasploit_process_tool]
-    agent = create_tool_calling_agent(
-        llm=llm,
-        prompt=prompt,
-        tools=tools
-    )
-
-    return AgentExecutor(
-        agent=agent, 
-        tools=tools, 
-        verbose=True,
-        callbacks=debug_callbacks
-    ), parser
 
 # Get agent executor and parser for the selected model
 if model_changed:
@@ -1002,177 +70,12 @@ if model_changed:
 
 agent_executor, parser = get_agent_executor(model_option)
 
-# Show which model is active
-st.sidebar.success(f"Active model: {model_option}")
+# Setup chat interface
+setup_chat()
 
-# Display existing messages
-for message in st.session_state.messages:
-    with st.chat_message(message["role"]):
-        st.write(message["content"])
-
-# Helper function to run agent with cancellation checks
-def run_with_cancellation_checks(agent_executor, inputs):
-    """Run the agent executor with periodic cancellation checks and error handling"""
-    # Ensure chat_history is included in inputs
-    if "chat_history" not in inputs:
-        inputs["chat_history"] = ""
-        
-    # Start execution in a separate thread
-    with ThreadPoolExecutor() as executor:
-        future = executor.submit(agent_executor.invoke, inputs)
-
-        # Check for cancellation while waiting for result
-        while not future.done():
-            if st.session_state.cancel_requested:
-                # We can't truly cancel the thread, but we can stop processing the result
-                return {"output": "Processing cancelled by user"}
-            time.sleep(0.1)  # Check every 100ms
-
-        try:
-            return future.result()
-        except Exception as e:
-            logger.error(f"Error in agent execution: {e}", exc_info=True)
-            # Log detailed information about the error
-            if "ValidationError" in str(e):
-                logger.error(f"Validation error details: {str(e)}")
-                # Try to extract more information about what was passed to the tool
-                if "NmapResultsInput" in str(e):
-                    logger.error("Error related to NmapResultsInput validation")
-                    if "input_value" in str(e):
-                        match = re.search(r"input_value=(\{.*?\})", str(e))
-                        if match:
-                            logger.error(f"Input value that caused error: {match.group(1)}")
-            
-            # Re-raise the exception to be handled by the outer try/except
-            raise
-
-# Get user input
+# Process user query if provided
 query = st.chat_input("Use NLP to input the target. Example: Scan a target IP 127.0.0.1")
-
-# Only process if there's a query
 if query:
-    # Add user message to chat history
-    st.session_state.messages.append({"role": "user", "content": query})
-
-    # Display user message
-    with st.chat_message("user"):
-        st.write(query)
-
-    # Set processing flag
-    st.session_state.is_processing = True
-    st.session_state.cancel_requested = False
-
-    # Create a container for the cancel button with a more prominent style
-    cancel_container = st.container()
-
-    # Show cancel button while processing with a more prominent style
-    if cancel_container.button("⚠️ Cancel Processing", key="cancel_button",
-                              use_container_width=True,
-                              type="primary"):
-        st.session_state.cancel_requested = True
-        st.warning("Cancellation requested. Waiting for current operation to complete...")
-        st.rerun()  # Rerun to update UI
-
-    # Check if cancellation was already requested
-    if st.session_state.cancel_requested:
-        st.warning("Processing cancelled by user")
-        st.session_state.is_processing = False
-    else:
-        # Show a spinner while processing
-        with st.spinner("Processing..."):
-            # Run the agent with cancellation checks
-            raw_response = run_with_cancellation_checks(agent_executor, {"query": query})
-
-            # Check if cancellation was requested during execution
-            if st.session_state.cancel_requested or raw_response.get("output") == "Processing cancelled by user":
-                st.warning("Processing cancelled by user")
-                st.session_state.is_processing = False
-                st.session_state.cancel_requested = False
-                st.rerun()  # Rerun to update UI
-            else:
-                try:
-                    # Try the Claude format first
-                    try:
-                        structured_response = parser.parse(raw_response.get("output")[0]["text"])
-                    except (KeyError, TypeError, IndexError):
-                        # Try alternative formats for OpenAI
-                        if isinstance(raw_response.get("output"), str):
-                            # If output is directly a string
-                            structured_response = parser.parse(raw_response.get("output"))
-                        elif hasattr(raw_response.get("output"), "content"):
-                            # If output has a content attribute
-                            structured_response = parser.parse(raw_response.get("output").content)
-                        else:
-                            # Try to extract content from the response
-                            output = raw_response.get("output")
-                            if isinstance(output, dict) and "content" in output:
-                                structured_response = parser.parse(output["content"])
-                            else:
-                                # Last resort - convert the whole output to string
-                                structured_response = parser.parse(str(output))
-
-                    # Add assistant response to chat history
-                    st.session_state.messages.append({"role": "assistant", "content": structured_response})
-
-                    # Display assistant response
-                    with st.chat_message("assistant"):
-                        # Display the main response
-                        st.write(structured_response)
-
-                        # If there are scan results, display them in a dedicated section
-                        if hasattr(structured_response, "scan_results") and structured_response.scan_results:
-                            with st.expander("Nmap Scan Results", expanded=True):
-                                st.code(structured_response.scan_results, language="bash")
-
-                        # If there are vulnerabilities, display them in a dedicated section
-                        if hasattr(structured_response, "vulnerabilities") and structured_response.vulnerabilities:
-                            with st.expander("Vulnerabilities Found", expanded=True):
-                                st.json(structured_response.vulnerabilities)
-
-                        # If there is an attack plan, display it in a dedicated section
-                        if hasattr(structured_response, "attack_plan") and structured_response.attack_plan:
-                            with st.expander("Attack Plan", expanded=True):
-                                # Display executive summary
-                                if "executive_summary" in structured_response.attack_plan:
-                                    st.subheader("Executive Summary")
-                                    st.write(structured_response.attack_plan["executive_summary"])
-
-                                # Display detailed plan
-                                if "detailed_plan" in structured_response.attack_plan:
-                                    st.subheader("Detailed Plan")
-                                    st.markdown(structured_response.attack_plan["detailed_plan"])
-
-                                # Display task list
-                                if "task_list" in structured_response.attack_plan:
-                                    st.subheader("Task List")
-                                    for task in structured_response.attack_plan["task_list"]:
-                                        st.markdown(f"**Task {task.get('id', '')}:** {task.get('name', '')}")
-                                        st.markdown(f"*Description:* {task.get('description', '')}")
-                                        if "commands" in task:
-                                            st.code("\n".join(task["commands"]), language="bash")
-                                        st.markdown("---")
-
-                        # If there are exploit paths, display them in a dedicated section
-                        if hasattr(structured_response, "exploit_paths") and structured_response.exploit_paths:
-                            with st.expander("Metasploit Exploit Paths", expanded=True):
-                                for exploit in structured_response.exploit_paths:
-                                    if "exploit_info" in exploit and "exploit_path" in exploit["exploit_info"]:
-                                        st.markdown(f"**Exploit for {exploit.get('vulnerability_name', '')}:**")
-                                        st.code(exploit["exploit_info"]["exploit_path"], language="bash")
-                                        if "exploit_description" in exploit["exploit_info"]:
-                                            st.markdown(f"*Description:* {exploit['exploit_info']['exploit_description']}")
-                                        st.markdown("---")
-
-                except Exception as e:
-                    error_msg = f"Error parsing response: {e}"
-                    st.error(error_msg)
-                    # Display raw response for debugging
-                    with st.expander("Raw Response (Debug)"):
-                        st.write(raw_response)
-                    print(error_msg, "Raw Response - ", raw_response)
-
-    # Reset processing flags after completion
-    st.session_state.is_processing = False
-    st.session_state.cancel_requested = False
+    process_query(query, agent_executor, parser)
 
 logger.info("Python script execution completed successfully")
