+++ python/models/model_handler.py
@@ -1,0 +1,162 @@
+"""
+Model handler for creating and managing LLM agents.
+"""
+
+import streamlit as st
+import sys
+import os
+from langchain.agents import AgentExecutor, create_tool_calling_agent
+from langchain_anthropic import ChatAnthropic
+from langchain_core.output_parsers import PydanticOutputParser
+from langchain_core.prompts import ChatPromptTemplate
+from langchain_openai import ChatOpenAI
+from pydantic import BaseModel, Field
+
+# Add the parent directory to sys.path to allow absolute imports
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+from callbacks.logging_callbacks import DetailedLoggingCallbackHandler
+from modules.metasploit_agent import metasploit_process_tool, metasploit_search_tool
+from modules.security_analyst import security_analyst_tool
+from modules.tools import nmap_tool, save_tool, search_tool, wiki_tool
+
+class ModelResponse(BaseModel):
+    """Model response schema for structured output parsing."""
+    topic: str
+    summary: str
+    sources: list[str]
+    tools_used: list[str]
+    scan_results: str = ""  # Optional field for nmap results
+    vulnerabilities: list = []  # List of vulnerabilities found
+    attack_plan: dict = {}  # Attack plan generated by security analyst
+    exploit_paths: list = []  # List of exploit paths found by metasploit agent
+    
+    @classmethod
+    def parse_obj(cls, obj):
+        """Custom parser to handle various input formats"""
+        try:
+            # If obj is a list, try to parse the first item
+            if isinstance(obj, list):
+                if obj and isinstance(obj[0], dict):
+                    return super().parse_obj(obj[0])
+                # If the list is empty or doesn't contain a dict, create a minimal valid response
+                return cls(
+                    topic="Error parsing response",
+                    summary="The response could not be parsed correctly.",
+                    sources=["N/A"],
+                    tools_used=["N/A"]
+                )
+            # If obj is a string, try to parse it as JSON
+            elif isinstance(obj, str):
+                try:
+                    import json
+                    parsed_obj = json.loads(obj)
+                    return cls.parse_obj(parsed_obj)
+                except json.JSONDecodeError:
+                    # If it's not valid JSON, create a minimal valid response
+                    return cls(
+                        topic="Error parsing response",
+                        summary=f"Could not parse response as JSON: {obj[:100]}...",
+                        sources=["N/A"],
+                        tools_used=["N/A"]
+                    )
+            # Otherwise, use the default parser
+            return super().parse_obj(obj)
+        except Exception as e:
+            # Fallback for any other parsing errors
+            import logging
+            logger = logging.getLogger(__name__)
+            logger.error(f"Error parsing ModelResponse: {e}")
+            return cls(
+                topic="Error parsing response",
+                summary=f"An error occurred while parsing the response: {str(e)}",
+                sources=["N/A"],
+                tools_used=["N/A"]
+            )
+
+@st.cache_resource(show_spinner=False)
+def get_agent_executor(model_name):
+    """Create and return an agent executor based on the selected model."""
+    if model_name == "Claude (Anthropic)":
+        llm = ChatAnthropic(model="claude-3-5-sonnet-20241022")
+    else:
+        llm = ChatOpenAI(temperature=0, model="gpt-4o-mini")
+
+    parser = PydanticOutputParser(pydantic_object=ModelResponse)
+    
+    # Create the callback handler
+    debug_callbacks = [DetailedLoggingCallbackHandler()]
+
+    prompt = ChatPromptTemplate.from_messages(
+        [
+            (
+                "system",
+                """
+                You are a senior cybersecurity research analyst who will help pentest a target system.
+                You have access to several tools:
+
+                1. scan_target (MCP) - Scan a target IP address or hostname with nmap. You can specify scan types:
+                   - basic: Standard TCP scan with version detection (default)
+                   - quick: Fast TCP scan of common ports
+                   - full: Comprehensive TCP scan of all ports with version detection
+                   - vuln: TCP scan with vulnerability scripts
+                   - udp_ports: TCP scan of common UDP service ports
+                   - version: Get nmap version information
+
+                   Example usage: scan_target(target="192.168.1.1", scan_type="vuln")
+
+                2. analyze_scan (MCP) - Analyze the most recent scan results to identify vulnerabilities, map them to MITRE ATT&CK framework, and generate an attack plan.
+                   IMPORTANT: This tool must only be used after running scan_target.
+                   Example usage: analyze_scan()
+
+                3. metasploit_search - Search for Metasploit exploits related to a vulnerability. Input can be a vulnerability name or a JSON object with vulnerability details.
+
+                4. metasploit_process_latest - Process the latest vulnerabilities from the database and search for Metasploit exploits for each one.
+
+                5. search - Search the web for cybersecurity information
+
+                6. wiki_tool - Query Wikipedia for relevant information
+
+                7. save_text_to_file - Save your findings to a file
+
+                When scanning targets, you MUST ALWAYS include the COMPLETE and UNMODIFIED nmap scan results in your response.
+                Never summarize or omit any part of the nmap scan output - the full raw output must be preserved.
+
+                Be thorough in your analysis of scan results, identifying potential vulnerabilities.
+
+                CRITICAL INSTRUCTION FOR SECURITY_ANALYST TOOL:
+                When a vulnerability is found (indicated by "State: VULNERABLE" in nmap results), use the security_analyst tool to analyze it, but you MUST ALWAYS provide the nmap_data parameter with the scan results. 
+                NEVER call security_analyst with an empty object {{}} or without the nmap_data parameter.
+                
+                CORRECT USAGE: security_analyst(nmap_data=scan_results)
+                INCORRECT USAGE: security_analyst() or security_analyst({{}})
+                
+                After analysis, use the metasploit_search tool to find relevant exploits.
+
+                IMPORTANT: Always set the scan_results field in your response to the complete nmap output.
+                If you use the security_analyst tool, use the raw_scan_output field from its response as your scan_results.
+
+                Previous conversation: {chat_history}
+
+                Answer the user query and use necessary tools.
+                Wrap the output in this format and provide no other text\n
+                {format_instructions}
+                """,
+            ),
+            ("human", "{query}"),
+            ("ai", "{agent_scratchpad}"),
+        ]
+    ).partial(format_instructions=parser.get_format_instructions())
+
+    tools = [search_tool, wiki_tool, save_tool, nmap_tool, security_analyst_tool, metasploit_search_tool, metasploit_process_tool]
+    agent = create_tool_calling_agent(
+        llm=llm,
+        prompt=prompt,
+        tools=tools
+    )
+
+    return AgentExecutor(
+        agent=agent, 
+        tools=tools, 
+        verbose=True,
+        callbacks=debug_callbacks
+    ), parser
