+++ python/modules/agents/mitre_attack_mapper_agent.py
@@ -38,6 +38,20 @@ class MitreAttackMapperAgent(BaseAgent):
             "tactic": "Impact",
             "description": "Attackers use the Slowloris technique to perform a denial of service attack by holding open multiple connections to the target web server with partial HTTP requests.",
             "mitigation": "Implement request timeout, limit concurrent connections per client, use reverse proxy with DoS protection."
+        },
+        "cve-2017-1001000": {
+            "technique_id": "T1068",
+            "technique_name": "Exploitation for Privilege Escalation",
+            "tactic": "Privilege Escalation",
+            "description": "Attackers exploit the WordPress REST API vulnerability (CVE-2017-1001000) that allows unauthenticated users to inject content into posts, leading to privilege escalation.",
+            "mitigation": "Update WordPress to version 4.7.2 or later, implement web application firewall rules to block malicious requests."
+        },
+        "http-vuln-cve2017-1001000": {
+            "technique_id": "T1068",
+            "technique_name": "Exploitation for Privilege Escalation",
+            "tactic": "Privilege Escalation",
+            "description": "Attackers exploit the WordPress REST API vulnerability (CVE-2017-1001000) that allows unauthenticated users to inject content into posts, leading to privilege escalation.",
+            "mitigation": "Update WordPress to version 4.7.2 or later, implement web application firewall rules to block malicious requests."
         }
     }
