+++ python/modules/security_analyst.py
@@ -119,8 +119,9 @@ class NmapResultsInput(BaseModel):
         description="Nmap scan results, either as a JSON object or raw text output from a previous nmap scan"
     )
     
-    class Config:
-        arbitrary_types_allowed = True
+    model_config = {
+        "arbitrary_types_allowed": True
+    }
 
 # Create the tool using StructuredTool
 security_analyst_tool = StructuredTool.from_function(
