+++ python/pyproject.toml
@@ -30,6 +30,7 @@ python_functions = "test_*"
 markers = [
     "unit: marks tests as unit tests",
     "integration: marks tests as integration tests",
+    "functional: marks tests as functional tests",
     "vulnerability_detection: tests for vulnerability detection",
     "mitre_mapping: tests for MITRE ATT&CK mapping",
     "attack_plan: tests for attack plan generation",
