+++ python/tests/README.md
@@ -1,94 +1,114 @@
-# Tests Directory
+# Test Framework
 
-This directory contains all tests for the pentest-agent-system Python module.
+This directory contains the test framework for the pentest-agent-system. The tests are organized into different categories based on their scope and purpose.
 
 ## Test Structure
 
-The tests are organized using pytest with fixtures and parameterization:
+The test framework is organized as follows:
 
 ```
 tests/
-├── conftest.py                  # Global fixtures and configuration
-├── data/                        # Test data directory
+├── conftest.py             # Common pytest fixtures and configuration
+├── data/                   # Test data and mock functions
 │   ├── __init__.py
-│   └── nmap_samples.py          # Sample nmap data in different formats
-├── unit/                        # Unit tests
+│   ├── mock_functions.py   # Mock functions for testing
+│   └── nmap_samples.py     # Sample nmap data for testing
+├── functional/             # Functional tests (end-to-end workflows)
 │   ├── __init__.py
-│   └── test_vulnerability_extractor.py
-└── integration/                 # Integration tests
-    ├── __init__.py
-    └── test_security_analyst.py
+│   └── test_end_to_end.py  # End-to-end functional tests
+├── integration/            # Integration tests (component interactions)
+│   ├── __init__.py
+│   └── test_security_analyst_integration.py  # Security analyst integration tests
+├── unit/                   # Unit tests (individual components)
+│   ├── __init__.py
+│   ├── test_attack_plan_generator.py  # Attack plan generator unit tests
+│   ├── test_mitre_mapper.py           # MITRE mapper unit tests
+│   └── test_vulnerability_extractor.py  # Vulnerability extractor unit tests
+└── sample_nmap_xml_output.txt  # Sample nmap output for testing
 ```
+
+## Test Categories
+
+### Unit Tests
+
+Unit tests focus on testing individual components in isolation. They are located in the `unit/` directory and are marked with the `@pytest.mark.unit` decorator.
+
+### Integration Tests
+
+Integration tests focus on testing the interactions between different components. They are located in the `integration/` directory and are marked with the `@pytest.mark.integration` decorator.
+
+### Functional Tests
+
+Functional tests focus on testing complete workflows from end to end. They are located in the `functional/` directory and are marked with the `@pytest.mark.functional` decorator.
 
 ## Running Tests
 
-Tests can be run using pytest:
+### Running All Tests
+
+To run all tests, use the following command:
 
 ```bash
-# Activate the virtual environment
 cd python
-source ./.venv/bin/activate
-
-# Run all tests
-python -m pytest
-
-# Run specific test categories
-python -m pytest tests/unit
-python -m pytest tests/integration
-
-# Run specific test files
-python -m pytest tests/unit/test_vulnerability_extractor.py
-python -m pytest tests/integration/test_security_analyst.py
-
-# Run tests with verbose output
-python -m pytest -v
-
-# Run tests with specific markers
-python -m pytest -m unit
-python -m pytest -m integration
-python -m pytest -m vulnerability_detection
+source .venv/bin/activate
+python -m pytest tests/
 ```
 
-## Test Markers
+### Running Specific Test Categories
 
-The following markers are available:
+To run only unit tests:
 
-- `unit`: Unit tests
-- `integration`: Integration tests
-- `vulnerability_detection`: Tests for vulnerability detection
-- `mitre_mapping`: Tests for MITRE ATT&CK mapping
-- `attack_plan`: Tests for attack plan generation
+```bash
+python -m pytest tests/ -m unit
+```
 
-## Fixtures
+To run only integration tests:
 
-Common fixtures are defined in `conftest.py`:
+```bash
+python -m pytest tests/ -m integration
+```
 
-- `configure_logging`: Configure logging for tests
-- `mock_db_manager`: Mock database manager
-- `vulnerability_extractor`: Vulnerability extractor agent
-- `security_analyst_orchestrator`: Security analyst orchestrator agent
-- Sample data fixtures: `ms17_010_data`, `multiple_vulnerabilities_data`, etc.
-- Expected results fixtures: `ms17_010_expected_vulns`, `ms17_010_expected_mitre`, etc.
+To run only functional tests:
 
-## Guidelines
+```bash
+python -m pytest tests/ -m functional
+```
 
-1. **All tests should be placed in this directory**
-   - Unit tests in `tests/unit/`
-   - Integration tests in `tests/integration/`
-   - Follow the naming convention `test_*.py` for test files
+### Running Specific Test Files
 
-2. **Test Data**
-   - Test data should be placed in `tests/data/`
-   - Use fixtures to access test data
+To run a specific test file:
 
-3. **Imports**
-   - Import fixtures from `conftest.py`
-   - Use parameterization for testing multiple scenarios
+```bash
+python -m pytest tests/unit/test_vulnerability_extractor.py
+```
 
-4. **Mock Functions**
-   - Use the mock functions in this directory (e.g., `mock_search.py`) for testing
-   - Import mock functions using relative imports:
-     ```python
-     from .mock_search import mock_search
-     ```
-   - Or use fixtures defined in `conftest.py`
+### Running Specific Test Functions
+
+To run a specific test function:
+
+```bash
+python -m pytest tests/unit/test_vulnerability_extractor.py::TestVulnerabilityExtractor::test_vulnerability_detection_dict
+```
+
+## Test Fixtures
+
+Common test fixtures are defined in `conftest.py`. These fixtures provide common functionality and test data for all tests.
+
+### Key Fixtures
+
+- `configure_logging`: Configures logging for tests
+- `mock_db_manager`: Provides a mock database manager for testing
+- `vulnerability_extractor`, `mitre_mapper`, `attack_plan_generator`, `security_analyst_orchestrator`: Provide instances of the respective agents for testing
+- `sample_nmap_data`: Provides sample nmap data from file
+- `ms17_010_data`, `multiple_vulnerabilities_data`, `port_script_vulnerabilities_data`: Provide specific vulnerability data for testing
+- `mock_search_function`: Provides a mock search function for testing
+
+## Adding New Tests
+
+When adding new tests, follow these guidelines:
+
+1. Place unit tests in the `unit/` directory
+2. Place integration tests in the `integration/` directory
+3. Place functional tests in the `functional/` directory
+4. Use appropriate pytest markers to categorize tests
+5. Use fixtures from `conftest.py` where possible
+6. Follow the naming convention: `test_*.py` for test files, `Test*` for test classes, and `test_*` for test functions
