+++ python/tests/conftest.py
@@ -1,5 +1,7 @@
 """
 Pytest configuration and fixtures.
+
+This module provides fixtures for all test modules in the project.
 """
 
 import pytest
@@ -7,27 +9,50 @@ import logging
 import json
 import sys
 import os
+import warnings
 from pathlib import Path
+
+# Filter out specific deprecation warnings
+warnings.filterwarnings("ignore", message="Failing to pass a value to the 'type_params' parameter of 'typing.ForwardRef._evaluate'", category=DeprecationWarning)
+warnings.filterwarnings("ignore", message="Support for class-based `config` is deprecated", category=DeprecationWarning)
 
 # Add the parent directory to the Python path for relative imports
 sys.path.append(str(Path(__file__).resolve().parent.parent))
 
 # Import the necessary modules
 from modules.agents.vulnerability_extractor_agent import VulnerabilityExtractorAgent
+from modules.agents.mitre_attack_mapper_agent import MitreAttackMapperAgent
+from modules.agents.attack_plan_generator_agent import AttackPlanGeneratorAgent
 from modules.agents.security_analyst_orchestrator_agent import SecurityAnalystOrchestratorAgent
+from modules.security_analyst import security_analyst_tool, analyze_nmap_results
+from modules.security_analyst_agent import analyze_scan_results, create_security_analyst_agent
 from .data.nmap_samples import (
     MS17_010_DATA, MULTIPLE_VULNERABILITIES_DATA, PORT_SCRIPT_VULNERABILITIES_DATA,
     MS17_010_DATA_STR, MULTIPLE_VULNERABILITIES_DATA_STR, PORT_SCRIPT_VULNERABILITIES_DATA_STR,
     MS17_010_EXPECTED_VULNS, MULTIPLE_VULNERABILITIES_EXPECTED_VULNS, PORT_SCRIPT_EXPECTED_VULNS,
     MS17_010_EXPECTED_MITRE, MULTIPLE_VULNERABILITIES_EXPECTED_MITRE, PORT_SCRIPT_EXPECTED_MITRE
 )
-from .mock_search import mock_search
+from .data.mock_functions import mock_search
 
 # Configure logging for tests
 @pytest.fixture(scope="session")
 def configure_logging():
     """Configure logging for tests."""
-    logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+    # Set up logging
+    logs_dir = Path(__file__).resolve().parent.parent.parent / "logs"
+    if not logs_dir.exists():
+        logs_dir.mkdir(parents=True, exist_ok=True)
+    
+    # Set up logging to file and console
+    log_file = logs_dir / 'pytest.log'
+    logging.basicConfig(
+        level=logging.INFO,
+        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
+        handlers=[
+            logging.FileHandler(str(log_file), mode='a'),
+            logging.StreamHandler(sys.stdout)
+        ]
+    )
     return logging.getLogger("test")
 
 # Mock database manager
@@ -50,20 +75,53 @@ def mock_db_manager():
         def store_attack_plan(self, scan_id, executive_summary, detailed_plan, task_list):
             logging.info(f"Storing attack plan: {executive_summary[:50]}...")
             return "mock_plan_id"
+        
+        def get_scan_results(self, scan_id):
+            logging.info(f"Getting scan results for scan_id: {scan_id}")
+            return {"target": "mock_target", "scan_type": "mock_scan_type", "raw_data": "mock_raw_data"}
+        
+        def get_vulnerabilities(self, scan_id):
+            logging.info(f"Getting vulnerabilities for scan_id: {scan_id}")
+            return [{"name": "mock_vuln", "description": "mock_description", "severity": "high"}]
     
     return MockDatabaseManager()
 
-# Vulnerability extractor agent
+# Agent fixtures
 @pytest.fixture
 def vulnerability_extractor(mock_db_manager, configure_logging):
     """Vulnerability extractor agent for testing."""
     return VulnerabilityExtractorAgent(db_manager=mock_db_manager, logger=configure_logging)
 
-# Security analyst orchestrator agent
+@pytest.fixture
+def mitre_mapper(mock_db_manager, configure_logging):
+    """MITRE ATT&CK mapper agent for testing."""
+    return MitreAttackMapperAgent(db_manager=mock_db_manager, logger=configure_logging)
+
+@pytest.fixture
+def attack_plan_generator(mock_db_manager, configure_logging):
+    """Attack plan generator agent for testing."""
+    return AttackPlanGeneratorAgent(db_manager=mock_db_manager, logger=configure_logging)
+
 @pytest.fixture
 def security_analyst_orchestrator(mock_db_manager, configure_logging):
     """Security analyst orchestrator agent for testing."""
     return SecurityAnalystOrchestratorAgent(db_manager=mock_db_manager, logger=configure_logging)
+
+@pytest.fixture
+def security_analyst_agent():
+    """Security analyst agent for testing."""
+    return create_security_analyst_agent()
+
+# Sample data fixtures
+@pytest.fixture
+def sample_nmap_data():
+    """Sample nmap data from file."""
+    sample_file = Path(__file__).resolve().parent / "sample_nmap_xml_output.txt"
+    if not sample_file.exists():
+        pytest.skip(f"Sample file not found: {sample_file}")
+    
+    with open(sample_file, "r") as f:
+        return f.read()
 
 # Sample nmap data fixtures
 @pytest.fixture
