+++ python/tests/data/__init__.py
@@ -1,0 +1,7 @@
+"""
+Test data package.
+"""
+
+from .mock_functions import mock_search
+
+__all__ = ["mock_search"]
