+++ python/tests/data/mock_functions.py
@@ -1,0 +1,44 @@
+"""
+Mock functions for testing.
+"""
+
+def mock_search(query):
+    """
+    Mock search function that returns predefined results based on the query.
+    
+    Args:
+        query: Search query
+        
+    Returns:
+        Predefined search results
+    """
+    if "ms17-010" in query or "eternalblue" in query:
+        return """
+        MITRE ATT&CK T1190 - Exploit Public-Facing Application (Initial Access)
+        
+        The MS17-010 vulnerability, also known as EternalBlue, is a critical remote code execution vulnerability
+        in Microsoft's SMBv1 server. It was exploited by the WannaCry ransomware in 2017.
+        
+        Attackers can exploit this vulnerability to gain remote code execution on vulnerable Windows systems.
+        
+        Mitigation strategies include:
+        - Apply security patches (MS17-010)
+        - Disable SMBv1
+        - Use firewalls to restrict access to SMB ports (139, 445)
+        """
+    elif "slowloris" in query:
+        return """
+        MITRE ATT&CK T1499 - Endpoint Denial of Service (Impact)
+        
+        The Slowloris attack is a denial of service attack that allows a single machine to take down a server
+        with minimal bandwidth by keeping many connections open and holding them open as long as possible.
+        
+        Attackers can use this technique to make a web server unavailable to legitimate users.
+        
+        Mitigation strategies include:
+        - Implement request timeout
+        - Limit concurrent connections per client
+        - Use reverse proxy with DoS protection
+        """
+    else:
+        return f"No specific MITRE ATT&CK information found for: {query}"
