+++ python/tests/functional/test_end_to_end.py
@@ -1,0 +1,161 @@
+#!/usr/bin/env python3
+"""
+End-to-end functional tests for the pentest agent system.
+
+This module contains functional tests that test the complete workflow of the system,
+from scanning to analysis and reporting.
+"""
+
+import json
+import logging
+import os
+import sys
+import pytest
+from pathlib import Path
+
+# Add the parent directory to the Python path for relative imports
+sys.path.append(str(Path(__file__).resolve().parent.parent.parent))
+
+# Set up logging
+logs_dir = Path(__file__).resolve().parent.parent.parent.parent / "logs"
+if not logs_dir.exists():
+    logs_dir.mkdir(parents=True, exist_ok=True)
+
+# Set up logging to file and console
+log_file = logs_dir / 'test_end_to_end.log'
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
+    handlers=[
+        logging.FileHandler(str(log_file), mode='a'),
+        logging.StreamHandler(sys.stdout)
+    ]
+)
+logger = logging.getLogger(__name__)
+
+# Import the necessary modules
+from modules.security_analyst import security_analyst_tool, analyze_nmap_results
+from modules.security_analyst_agent import analyze_scan_results, create_security_analyst_agent
+from tests.data.mock_functions import mock_search
+
+@pytest.mark.functional
+class TestEndToEnd:
+    """End-to-end functional tests for the pentest agent system."""
+    
+    def test_security_analyst_tool_with_sample_data(self):
+        """Test the security_analyst_tool with sample nmap data."""
+        logger.info("Testing security_analyst_tool with sample nmap data")
+        
+        # Load sample nmap data
+        sample_file = Path(__file__).resolve().parent.parent / "sample_nmap_xml_output.txt"
+        assert sample_file.exists(), f"Sample file not found: {sample_file}"
+            
+        with open(sample_file, "r") as f:
+            sample_data = f.read()
+        
+        # Run the analyze_nmap_results function with mock search
+        result = analyze_nmap_results(sample_data, custom_search_func=mock_search)
+        
+        # Verify the result contains expected fields
+        assert "attack_plan" in result, "No attack plan found in result"
+        assert "raw_scan_output" in result, "No raw scan output found in result"
+        assert "target" in result, "No target found in result"
+        
+        # Check if vulnerabilities were found
+        if "vulnerabilities" in result and len(result["vulnerabilities"]) > 0:
+            # If vulnerabilities were found, verify they have the expected structure
+            for vuln in result["vulnerabilities"]:
+                assert "name" in vuln, "Vulnerability missing name"
+                assert "description" in vuln, "Vulnerability missing description"
+                assert "raw_data" in vuln, "Vulnerability missing raw_data"
+            
+            # Verify MITRE mappings
+            assert "mitre_mappings" in result, "No MITRE mappings found in result"
+            for mapping in result["mitre_mappings"]:
+                assert "technique_id" in mapping, "MITRE mapping missing technique_id"
+                assert "technique_name" in mapping, "MITRE mapping missing technique_name"
+                assert "tactic" in mapping, "MITRE mapping missing tactic"
+        else:
+            # If no vulnerabilities were found, log a warning but don't fail the test
+            logger.warning("No vulnerabilities found in the sample data. This might be expected if the vulnerability extractor doesn't support the XML format.")
+        
+        # Log results safely
+        vulnerabilities = result.get('vulnerabilities', [])
+        mitre_mappings = result.get('mitre_mappings', [])
+        logger.info(f"Found {len(vulnerabilities)} vulnerabilities")
+        logger.info(f"Found {len(mitre_mappings)} MITRE mappings")
+        logger.info(f"Attack plan executive summary: {result['attack_plan'].get('executive_summary')}")
+    
+    def test_security_analyst_agent_with_sample_data(self):
+        """Test the security_analyst_agent with sample nmap data."""
+        logger.info("Testing security_analyst_agent with sample nmap data")
+        
+        # Load sample nmap data
+        sample_file = Path(__file__).resolve().parent.parent / "sample_nmap_xml_output.txt"
+        assert sample_file.exists(), f"Sample file not found: {sample_file}"
+            
+        with open(sample_file, "r") as f:
+            sample_data = f.read()
+        
+        # Skip this test if OpenAI API key is not set
+        openai_api_key = os.environ.get("OPENAI_API_KEY")
+        if not openai_api_key:
+            pytest.skip("OPENAI_API_KEY environment variable not set")
+            
+        # Analyze the scan results
+        try:
+            result = analyze_scan_results(sample_data, model_name="gpt-4o-mini", save_to_db=False)
+            
+            # Verify the result
+            assert "output" in result, "No output found in result"
+            output = result["output"]
+            if isinstance(output, dict):
+                if "vulnerabilities" in output and len(output["vulnerabilities"]) > 0:
+                    logger.info(f"Found {len(output['vulnerabilities'])} vulnerabilities")
+                else:
+                    logger.warning("No vulnerabilities found in the output. This might be expected if the vulnerability extractor doesn't support the XML format.")
+            else:
+                logger.info(f"Output: {output}")
+        except Exception as e:
+            logger.warning(f"Error analyzing scan results: {e}")
+            pytest.skip(f"Error analyzing scan results: {e}")
+    
+    @pytest.mark.skip(reason="This test runs a real nmap scan and should be run manually")
+    def test_scan_and_analyze(self):
+        """Test running a scan and analyzing the results in one step."""
+        logger.info("Testing security analyst agent with direct scanning")
+        
+        # Create the agent
+        agent = create_security_analyst_agent()
+        
+        # Run a scan and analyze the results
+        target = "127.0.0.1"  # Use localhost for testing
+        scan_type = "quick"   # Use quick scan for faster results
+        
+        logger.info(f"Running nmap scan on {target} with scan type {scan_type}")
+        result = agent.invoke({
+            "input": f"Run an nmap scan against {target} with scan type {scan_type} and analyze the results"
+        })
+        
+        # Verify the result
+        assert "output" in result, "No output found in result"
+        output = result["output"]
+        if isinstance(output, dict):
+            logger.info(f"Scan ID: {output.get('scan_id')}")
+            logger.info(f"Target: {output.get('target')}")
+            logger.info(f"Vulnerabilities found: {output.get('vulnerabilities_found')}")
+        else:
+            logger.info(f"Output: {output}")
+
+if __name__ == "__main__":
+    # Create a test instance
+    test = TestEndToEnd()
+    
+    # Run the tests
+    print("Running test_security_analyst_tool_with_sample_data...")
+    test.test_security_analyst_tool_with_sample_data()
+    
+    print("\nRunning test_security_analyst_agent_with_sample_data...")
+    test.test_security_analyst_agent_with_sample_data()
+    
+    print("\nTests completed successfully!")
