+++ python/tests/integration/__init__.py
@@ -1,0 +1,3 @@
+"""
+Integration tests package.
+"""
