--- python/tests/sample_nmap_output.txt
@@ -1,47 +1,0 @@
-NMAP SCAN RESULTS (vuln scan of 192.168.1.100):
-==================================================
-Starting Nmap 7.93 ( https://nmap.org ) at 2025-03-22 13:56 EDT
-Nmap scan report for 192.168.1.100
-Host is up (0.0054s latency).
-Not shown: 995 closed tcp ports (conn-refused)
-PORT     STATE SERVICE       VERSION
-135/tcp  open  msrpc         Microsoft Windows RPC
-139/tcp  open  netbios-ssn   Microsoft Windows netbios-ssn
-445/tcp  open  microsoft-ds  Microsoft Windows 7 - 10 microsoft-ds (workgroup: WORKGROUP)
-3389/tcp open  ms-wbt-server Microsoft Terminal Services
-8080/tcp open  http          Apache Tomcat/Coyote JSP engine 1.1
-Service Info: Host: WIN-PC; OS: Windows; CPE: cpe:/o:microsoft:windows
-
-Host script results:
-| smb-vuln-ms17-010: 
-|   VULNERABLE:
-|   Remote Code Execution vulnerability in Microsoft SMBv1 servers (ms17-010)
-|     State: VULNERABLE
-|     IDs:  CVE:CVE-2017-0143
-|     Risk factor: HIGH
-|       A critical remote code execution vulnerability exists in Microsoft SMBv1
-|        servers (ms17-010).
-|           
-|     Disclosure date: 2017-03-14
-|     References:
-|       https://blogs.technet.microsoft.com/msrc/2017/05/12/customer-guidance-for-wannacrypt-attacks/
-|       https://technet.microsoft.com/en-us/library/security/ms17-010.aspx
-|_      https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2017-0143
-
-| http-slowloris-check: 
-|   VULNERABLE:
-|   Slowloris DOS attack
-|     State: VULNERABLE
-|     IDs:  CVE:CVE-2007-6750
-|       Slowloris tries to keep many connections to the target web server open and hold
-|       them open as long as possible.  It accomplishes this by opening connections to
-|       the target web server and sending a partial request. By doing so, it starves
-|       the http server's resources causing Denial Of Service.
-|       
-|     Disclosure date: 2009-09-17
-|     References:
-|       https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2007-6750
-|_      http://ha.ckers.org/slowloris/
-
-Nmap done: 1 IP address (1 host up) scanned in 15.23 seconds
-==================================================
