+++ python/tests/test_agent_architecture.py
@@ -27,7 +27,7 @@ def test_individual_agents():
     db_manager = DatabaseManager()
     
     # Load the sample Nmap output
-    sample_file = Path(__file__).resolve().parent / "sample_nmap_output.txt"
+    sample_file = Path(__file__).resolve().parent / "sample_nmap_xml_output.txt"
     with open(sample_file, "r") as f:
         sample_data = f.read()
     
@@ -67,7 +67,7 @@ def test_orchestrator_agent():
     orchestrator = SecurityAnalystOrchestratorAgent(db_manager)
     
     # Load the sample Nmap output
-    sample_file = Path(__file__).resolve().parent / "sample_nmap_output.txt"
+    sample_file = Path(__file__).resolve().parent / "sample_nmap_xml_output.txt"
     with open(sample_file, "r") as f:
         sample_data = f.read()
     
@@ -115,7 +115,7 @@ def test_backward_compatibility():
     from modules.security_analyst import analyze_nmap_results
     
     # Load the sample Nmap output
-    sample_file = Path(__file__).resolve().parent / "sample_nmap_output.txt"
+    sample_file = Path(__file__).resolve().parent / "sample_nmap_xml_output.txt"
     with open(sample_file, "r") as f:
         sample_data = f.read()
