+++ python/tests/test_enhanced_analysis.py
@@ -16,7 +16,7 @@ from src.pentest_agent_system.result_schema import AnalysisResult
 def main():
     """Run a test of the enhanced security analyst functionality."""
     # Load the sample Nmap output
-    sample_file = Path(__file__).resolve().parent / "sample_nmap_output.txt"
+    sample_file = Path(__file__).resolve().parent / "sample_nmap_xml_output.txt"
     with open(sample_file, "r") as f:
         sample_data = f.read()
