+++ python/tests/test_sample_output.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """
-Test script for the security_analyst module using the sample_nmap_output.txt file.
+Test script for the security_analyst module using the sample_nmap_xml_output.txt file.
 """
 
 import sys
@@ -13,10 +13,10 @@ from modules.security_analyst import analyze_nmap_results
 
 def main():
     """Run the test"""
-    print("Testing security_analyst module with sample_nmap_output.txt...\n")
+    print("Testing security_analyst module with sample_nmap_xml_output.txt...\n")
     
     # Read the sample nmap output using Path for proper file location
-    sample_file = Path(__file__).resolve().parent / 'sample_nmap_output.txt'
+    sample_file = Path(__file__).resolve().parent / 'sample_nmap_xml_output.txt'
     with open(sample_file, 'r') as f:
         sample_output = f.read()
