+++ python/tests/test_security_analyst.py
@@ -73,7 +73,7 @@ def test_with_sample_data():
     logger.info("Testing security_analyst_tool with sample nmap data from file")
     try:
         # Update the path to the sample file since we're now in the tests directory
-        sample_file = Path(__file__).resolve().parent / "sample_nmap_output.txt"
+        sample_file = Path(__file__).resolve().parent / "sample_nmap_xml_output.txt"
         if not sample_file.exists():
             logger.error(f"Sample file not found: {sample_file}")
             return False
