+++ python/tests/test_security_analyst_agent.py
@@ -20,7 +20,7 @@ def test_with_file():
     print("Testing security analyst agent with sample nmap output file...")
     
     # Load sample nmap scan results
-    sample_file = Path(__file__).resolve().parent.parent / "sample_nmap_output.txt"
+    sample_file = Path(__file__).resolve().parent.parent / "sample_nmap_xml_output.txt"
     if not sample_file.exists():
         print(f"Error: Sample file not found: {sample_file}")
         return
