+++ python/tests/test_security_analyst_direct.py
@@ -39,7 +39,7 @@ def main():
     logger.info("Testing security_analyst_tool directly with sample nmap data")
     
     # Update the path to the sample file since we're now in the tests directory
-    sample_file = Path(__file__).resolve().parent.parent / "sample_nmap_output.txt"
+    sample_file = Path(__file__).resolve().parent.parent / "sample_nmap_xml_output.txt"
     if not sample_file.exists():
         logger.error(f"Sample file not found: {sample_file}")
         return
