+++ python/tests/test_security_analyst_mock.py
@@ -39,7 +39,7 @@ def main():
     logger.info("Testing security_analyst module with real nmap data and mock search")
     
     # Read sample nmap data
-    sample_file = Path(__file__).resolve().parent / "sample_nmap_output.txt"
+    sample_file = Path(__file__).resolve().parent / "sample_nmap_xml_output.txt"
     if not sample_file.exists():
         logger.error(f"Sample file not found: {sample_file}")
         return
