+++ python/tests/test_security_analyst_real.py
@@ -38,7 +38,7 @@ def main():
     logger.info("Testing security_analyst module with real nmap data")
     
     # Read sample nmap data
-    sample_file = Path(__file__).resolve().parent / "sample_nmap_output.txt"
+    sample_file = Path(__file__).resolve().parent / "sample_nmap_xml_output.txt"
     if not sample_file.exists():
         logger.error(f"Sample file not found: {sample_file}")
         return
