+++ python/tests/unit/test_attack_plan_generator.py
@@ -1,0 +1,180 @@
+"""
+Unit tests for the attack plan generator agent.
+"""
+
+import pytest
+import logging
+from modules.agents.attack_plan_generator_agent import AttackPlanGeneratorAgent
+
+@pytest.mark.unit
+@pytest.mark.attack_plan
+class TestAttackPlanGenerator:
+    """Tests for the attack plan generator agent."""
+    
+    def test_generate_plan_ms17_010(self, attack_plan_generator, ms17_010_data, ms17_010_expected_mitre):
+        """Test generating attack plan for MS17-010 vulnerability."""
+        # Add target to the scan results
+        scan_results = {
+            "target": "10.0.255.165",
+            "scan_type": "vulnerability",
+            "scan_id": "test_scan_id",
+            "vulnerabilities": [{"id": "vuln-001", "name": "smb-vuln-ms17-010", "description": "MS17-010 vulnerability", "raw_data": "VULNERABLE"}]
+        }
+        
+        # Create proper MITRE mappings
+        mitre_mappings = [
+            {
+                "technique_id": "T1190",
+                "technique_name": "Exploit Public-Facing Application",
+                "tactic": "Initial Access",
+                "description": "Attackers may attempt to exploit vulnerabilities in public-facing applications.",
+                "mitigation": "Keep systems patched and updated."
+            }
+        ]
+        
+        # Generate attack plan
+        attack_plan = attack_plan_generator.generate_plan(scan_results, mitre_mappings)
+        
+        # Verify plan structure
+        assert "executive_summary" in attack_plan, "Attack plan missing executive_summary"
+        assert "detailed_plan" in attack_plan, "Attack plan missing detailed_plan"
+        assert "task_list" in attack_plan, "Attack plan missing task_list"
+        
+        # Verify executive summary
+        assert len(attack_plan["executive_summary"]) > 0, "Executive summary is empty"
+        
+        # Verify detailed plan
+        assert len(attack_plan["detailed_plan"]) > 0, "Detailed plan is empty"
+        
+        # Verify task list
+        assert len(attack_plan["task_list"]) > 0, "Task list is empty"
+        
+        # Verify task structure
+        for task in attack_plan["task_list"]:
+            assert "id" in task, "Task missing id"
+            assert "name" in task, "Task missing name"
+            assert "description" in task, "Task missing description"
+            assert "commands" in task, "Task missing commands"
+    
+    def test_generate_plan_multiple_vulnerabilities(self, attack_plan_generator, multiple_vulnerabilities_data, multiple_vulnerabilities_expected_mitre):
+        """Test generating attack plan for multiple vulnerabilities."""
+        # Add target to the scan results
+        scan_results = {
+            "target": "192.168.1.100",
+            "scan_type": "vulnerability",
+            "scan_id": "test_scan_id",
+            "vulnerabilities": [
+                {"id": "vuln-001", "name": "smb-vuln-ms17-010", "description": "MS17-010 vulnerability", "raw_data": "VULNERABLE"},
+                {"id": "vuln-002", "name": "http-slowloris-check", "description": "Slowloris DOS attack", "raw_data": "VULNERABLE"}
+            ]
+        }
+        
+        # Create proper MITRE mappings
+        mitre_mappings = [
+            {
+                "technique_id": "T1190",
+                "technique_name": "Exploit Public-Facing Application",
+                "tactic": "Initial Access",
+                "description": "Attackers may attempt to exploit vulnerabilities in public-facing applications.",
+                "mitigation": "Keep systems patched and updated."
+            },
+            {
+                "technique_id": "T1499",
+                "technique_name": "Endpoint Denial of Service",
+                "tactic": "Impact",
+                "description": "Attackers may attempt to deny availability to services.",
+                "mitigation": "Implement rate limiting and DoS protection."
+            }
+        ]
+        
+        # Generate attack plan
+        attack_plan = attack_plan_generator.generate_plan(scan_results, mitre_mappings)
+        
+        # Verify plan structure
+        assert "executive_summary" in attack_plan, "Attack plan missing executive_summary"
+        assert "detailed_plan" in attack_plan, "Attack plan missing detailed_plan"
+        assert "task_list" in attack_plan, "Attack plan missing task_list"
+        
+        # Verify task list
+        assert len(attack_plan["task_list"]) > 0, "Task list is empty"
+        
+        # Verify that the plan addresses multiple vulnerabilities
+        summary = attack_plan["executive_summary"].lower()
+        assert "multiple" in summary or "several" in summary or "various" in summary or "2 vulnerabilities" in summary, "Executive summary does not address multiple vulnerabilities"
+    
+    def test_generate_plan_empty_data(self, attack_plan_generator):
+        """Test generating attack plan with empty data."""
+        # Generate attack plan with empty data but with target
+        scan_results = {
+            "target": "unknown",
+            "scan_type": "unknown",
+            "scan_id": "test_scan_id",
+            "vulnerabilities": []
+        }
+        attack_plan = attack_plan_generator.generate_plan(scan_results, [])
+        
+        # Verify plan structure
+        assert "executive_summary" in attack_plan, "Attack plan missing executive_summary"
+        assert "detailed_plan" in attack_plan, "Attack plan missing detailed_plan"
+        assert "task_list" in attack_plan, "Attack plan missing task_list"
+        
+        # Verify that the plan acknowledges the lack of vulnerabilities
+        summary = attack_plan["executive_summary"].lower()
+        assert "no vulnerabilities" in summary or "no significant" in summary, "Executive summary does not acknowledge lack of vulnerabilities"
+        
+        # Verify task list is empty or contains only general tasks
+        if len(attack_plan["task_list"]) > 0:
+            # If there are tasks, they should be general recommendations
+            for task in attack_plan["task_list"]:
+                assert "general" in task["description"].lower() or "recommend" in task["description"].lower(), "Task is not a general recommendation"
+    
+    def test_task_list_structure(self, attack_plan_generator, ms17_010_data, ms17_010_expected_mitre):
+        """Test the structure of the task list in the attack plan."""
+        # Add target to the scan results
+        scan_results = {
+            "target": "10.0.255.165",
+            "scan_type": "vulnerability",
+            "scan_id": "test_scan_id",
+            "vulnerabilities": [
+                {
+                    "id": "vuln-001",
+                    "name": "smb-vuln-ms17-010", 
+                    "description": "MS17-010 vulnerability", 
+                    "raw_data": "VULNERABLE"
+                }
+            ]
+        }
+        
+        # Create proper MITRE mappings
+        mitre_mappings = [
+            {
+                "technique_id": "T1190",
+                "technique_name": "Exploit Public-Facing Application",
+                "tactic": "Initial Access",
+                "description": "Attackers may attempt to exploit vulnerabilities in public-facing applications.",
+                "mitigation": "Keep systems patched and updated."
+            }
+        ]
+        
+        # Generate attack plan
+        attack_plan = attack_plan_generator.generate_plan(scan_results, mitre_mappings)
+        
+        # Verify task list
+        assert len(attack_plan["task_list"]) > 0, "Task list is empty"
+        
+        # Verify task structure
+        for task in attack_plan["task_list"]:
+            assert "id" in task, "Task missing id"
+            assert "name" in task, "Task missing name"
+            assert "description" in task, "Task missing description"
+            assert "commands" in task, "Task missing commands"
+            
+            # Verify id format
+            assert isinstance(task["id"], int) or task["id"].isdigit(), f"Invalid task id format: {task['id']}"
+            
+            # Verify name and description
+            assert len(task["name"]) > 0, "Task name is empty"
+            assert len(task["description"]) > 0, "Task description is empty"
+            
+            # Verify commands
+            assert isinstance(task["commands"], list), "Commands is not a list"
