+++ python/tests/unit/test_mitre_mapper.py
@@ -1,0 +1,95 @@
+"""
+Unit tests for the MITRE ATT&CK mapper agent.
+"""
+
+import pytest
+import logging
+from modules.agents.mitre_attack_mapper_agent import MitreAttackMapperAgent
+
+@pytest.mark.unit
+@pytest.mark.mitre_mapping
+class TestMitreMapper:
+    """Tests for the MITRE ATT&CK mapper agent."""
+    
+    def test_map_vulnerabilities_ms17_010(self, mitre_mapper, ms17_010_expected_vulns, ms17_010_expected_mitre):
+        """Test mapping MS17-010 vulnerability to MITRE ATT&CK."""
+        # Create vulnerability dictionaries
+        vulnerabilities = [
+            {"id": "vuln-001", "name": "smb-vuln-ms17-010", "description": "MS17-010 vulnerability", "raw_data": "VULNERABLE"}
+        ]
+        
+        # Map vulnerabilities to MITRE ATT&CK
+        mitre_mappings = mitre_mapper.map_vulnerabilities(vulnerabilities)
+        
+        # Verify mappings
+        assert len(mitre_mappings) > 0, "No MITRE mappings found"
+        
+        # Verify specific mappings
+        technique_ids = [m["technique_id"] for m in mitre_mappings]
+        assert "T1190" in technique_ids, "Expected technique T1190 not found"
+    
+    def test_map_vulnerabilities_multiple(self, mitre_mapper, multiple_vulnerabilities_expected_vulns, multiple_vulnerabilities_expected_mitre):
+        """Test mapping multiple vulnerabilities to MITRE ATT&CK."""
+        # Create vulnerability dictionaries
+        vulnerabilities = [
+            {"id": "vuln-001", "name": "smb-vuln-ms17-010", "description": "MS17-010 vulnerability", "raw_data": "VULNERABLE"},
+            {"id": "vuln-002", "name": "http-slowloris-check", "description": "Slowloris DOS attack", "raw_data": "VULNERABLE"}
+        ]
+        
+        # Map vulnerabilities to MITRE ATT&CK
+        mitre_mappings = mitre_mapper.map_vulnerabilities(vulnerabilities)
+        
+        # Verify mappings
+        assert len(mitre_mappings) > 0, "No MITRE mappings found"
+        
+        # Verify specific mappings
+        technique_ids = [m["technique_id"] for m in mitre_mappings]
+        assert "T1190" in technique_ids, "Expected technique T1190 not found"
+        assert "T1499" in technique_ids, "Expected technique T1499 not found"
+    
+    def test_map_vulnerabilities_port_script(self, mitre_mapper, port_script_expected_vulns, port_script_expected_mitre):
+        """Test mapping port script vulnerabilities to MITRE ATT&CK."""
+        # Create vulnerability dictionaries
+        vulnerabilities = [
+            {"id": "vuln-003", "name": "http-vuln-cve2017-1001000", "description": "Web vulnerability", "raw_data": "VULNERABLE"}
+        ]
+        
+        # Map vulnerabilities to MITRE ATT&CK
+        mitre_mappings = mitre_mapper.map_vulnerabilities(vulnerabilities)
+        
+        # Verify mappings
+        assert len(mitre_mappings) > 0, "No MITRE mappings found"
+        
+        # Verify specific mappings
+        technique_ids = [m["technique_id"] for m in mitre_mappings]
+        assert "T1068" in technique_ids, "Expected technique T1068 not found"
+    
+    def test_mapping_structure(self, mitre_mapper, ms17_010_expected_vulns):
+        """Test the structure of MITRE mappings."""
+        # Create vulnerability dictionaries
+        vulnerabilities = [
+            {"id": "vuln-001", "name": "smb-vuln-ms17-010", "description": "MS17-010 vulnerability", "raw_data": "VULNERABLE"}
+        ]
+        
+        # Map vulnerabilities to MITRE ATT&CK
+        mitre_mappings = mitre_mapper.map_vulnerabilities(vulnerabilities)
+        
+        # Verify mapping structure
+        for mapping in mitre_mappings:
+            assert "technique_id" in mapping, "Mapping missing technique_id"
+            assert "technique_name" in mapping, "Mapping missing technique_name"
+            assert "tactic" in mapping, "Mapping missing tactic"
+            assert "description" in mapping, "Mapping missing description"
+            assert "mitigation" in mapping, "Mapping missing mitigation"
+            
+            # Verify technique_id format (e.g., T1190)
+            assert mapping["technique_id"].startswith("T"), f"Invalid technique_id format: {mapping['technique_id']}"
+            assert len(mapping["technique_id"]) >= 5, f"Invalid technique_id length: {mapping['technique_id']}"
+    
+    def test_empty_vulnerabilities(self, mitre_mapper):
+        """Test mapping empty vulnerabilities list."""
+        # Map empty vulnerabilities list
+        mitre_mappings = mitre_mapper.map_vulnerabilities([])
+        
+        # Verify empty mappings
+        assert len(mitre_mappings) == 0, "Expected empty mappings for empty vulnerabilities"
