+++ python/ui/chat_interface.py
@@ -1,0 +1,207 @@
+"""
+Chat interface UI components for the Streamlit application.
+"""
+
+import streamlit as st
+import time
+import json
+import re
+import logging
+import sys
+import os
+from concurrent.futures import ThreadPoolExecutor
+
+# Add the parent directory to sys.path to allow absolute imports
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+logger = logging.getLogger(__name__)
+
+def setup_chat():
+    """Initialize the chat interface."""
+    # Initialize session state for chat history
+    if "messages" not in st.session_state:
+        st.session_state.messages = []
+
+    # Initialize session state variables for cancellation
+    if "is_processing" not in st.session_state:
+        st.session_state.is_processing = False
+    if "cancel_requested" not in st.session_state:
+        st.session_state.cancel_requested = False
+    
+    # Display existing messages
+    for message in st.session_state.messages:
+        with st.chat_message(message["role"]):
+            st.write(message["content"])
+
+def process_query(query, agent_executor, parser):
+    """Process a user query using the agent executor."""
+    # Add user message to chat history
+    st.session_state.messages.append({"role": "user", "content": query})
+
+    # Display user message
+    with st.chat_message("user"):
+        st.write(query)
+
+    # Set processing flag
+    st.session_state.is_processing = True
+    st.session_state.cancel_requested = False
+
+    # Create a container for the cancel button with a more prominent style
+    cancel_container = st.container()
+
+    # Show cancel button while processing with a more prominent style
+    if cancel_container.button("⚠️ Cancel Processing", key="cancel_button",
+                              use_container_width=True,
+                              type="primary"):
+        st.session_state.cancel_requested = True
+        st.warning("Cancellation requested. Waiting for current operation to complete...")
+        st.rerun()  # Rerun to update UI
+
+    # Check if cancellation was already requested
+    if st.session_state.cancel_requested:
+        st.warning("Processing cancelled by user")
+        st.session_state.is_processing = False
+    else:
+        # Show a spinner while processing
+        with st.spinner("Processing..."):
+            # Run the agent with cancellation checks
+            raw_response = run_with_cancellation_checks(agent_executor, {"query": query})
+
+            # Check if cancellation was requested during execution
+            if st.session_state.cancel_requested or raw_response.get("output") == "Processing cancelled by user":
+                st.warning("Processing cancelled by user")
+                st.session_state.is_processing = False
+                st.session_state.cancel_requested = False
+                st.rerun()  # Rerun to update UI
+            else:
+                process_agent_response(raw_response, parser)
+
+    # Reset processing flags after completion
+    st.session_state.is_processing = False
+    st.session_state.cancel_requested = False
+
+def run_with_cancellation_checks(agent_executor, inputs):
+    """Run the agent executor with periodic cancellation checks and error handling."""
+    # Ensure chat_history is included in inputs
+    if "chat_history" not in inputs:
+        inputs["chat_history"] = ""
+        
+    # Start execution in a separate thread
+    with ThreadPoolExecutor() as executor:
+        future = executor.submit(agent_executor.invoke, inputs)
+
+        # Check for cancellation while waiting for result
+        while not future.done():
+            if st.session_state.cancel_requested:
+                # We can't truly cancel the thread, but we can stop processing the result
+                return {"output": "Processing cancelled by user"}
+            time.sleep(0.1)  # Check every 100ms
+
+        try:
+            return future.result()
+        except Exception as e:
+            logger.error(f"Error in agent execution: {e}", exc_info=True)
+            # Log detailed information about the error
+            if "ValidationError" in str(e):
+                logger.error(f"Validation error details: {str(e)}")
+                # Try to extract more information about what was passed to the tool
+                if "NmapResultsInput" in str(e):
+                    logger.error("Error related to NmapResultsInput validation")
+                    if "input_value" in str(e):
+                        match = re.search(r"input_value=(\{.*?\})", str(e))
+                        if match:
+                            logger.error(f"Input value that caused error: {match.group(1)}")
+            
+            # Re-raise the exception to be handled by the outer try/except
+            raise
+
+def process_agent_response(raw_response, parser):
+    """Process and display the agent's response."""
+    try:
+        # Try the Claude format first
+        try:
+            structured_response = parser.parse(raw_response.get("output")[0]["text"])
+        except (KeyError, TypeError, IndexError):
+            # Try alternative formats for OpenAI
+            if isinstance(raw_response.get("output"), str):
+                # If output is directly a string
+                structured_response = parser.parse(raw_response.get("output"))
+            elif hasattr(raw_response.get("output"), "content"):
+                # If output has a content attribute
+                structured_response = parser.parse(raw_response.get("output").content)
+            else:
+                # Try to extract content from the response
+                output = raw_response.get("output")
+                if isinstance(output, dict) and "content" in output:
+                    structured_response = parser.parse(output["content"])
+                else:
+                    # Last resort - convert the whole output to string
+                    structured_response = parser.parse(str(output))
+
+        # Add assistant response to chat history
+        st.session_state.messages.append({"role": "assistant", "content": structured_response})
+
+        # Display assistant response
+        with st.chat_message("assistant"):
+            # Display the main response
+            st.write(structured_response)
+
+            # If there are scan results, display them in a dedicated section
+            if hasattr(structured_response, "scan_results") and structured_response.scan_results:
+                with st.expander("Nmap Scan Results", expanded=True):
+                    st.code(structured_response.scan_results, language="bash")
+
+            # If there are vulnerabilities, display them in a dedicated section
+            if hasattr(structured_response, "vulnerabilities") and structured_response.vulnerabilities:
+                with st.expander("Vulnerabilities Found", expanded=True):
+                    st.json(structured_response.vulnerabilities)
+
+            # If there is an attack plan, display it in a dedicated section
+            if hasattr(structured_response, "attack_plan") and structured_response.attack_plan:
+                display_attack_plan(structured_response.attack_plan)
+
+            # If there are exploit paths, display them in a dedicated section
+            if hasattr(structured_response, "exploit_paths") and structured_response.exploit_paths:
+                display_exploit_paths(structured_response.exploit_paths)
+
+    except Exception as e:
+        error_msg = f"Error parsing response: {e}"
+        st.error(error_msg)
+        # Display raw response for debugging
+        with st.expander("Raw Response (Debug)"):
+            st.write(raw_response)
+        print(error_msg, "Raw Response - ", raw_response)
+
+def display_attack_plan(attack_plan):
+    """Display the attack plan in a structured format."""
+    with st.expander("Attack Plan", expanded=True):
+        # Display executive summary
+        if "executive_summary" in attack_plan:
+            st.subheader("Executive Summary")
+            st.write(attack_plan["executive_summary"])
+
+        # Display detailed plan
+        if "detailed_plan" in attack_plan:
+            st.subheader("Detailed Plan")
+            st.markdown(attack_plan["detailed_plan"])
+
+        # Display task list
+        if "task_list" in attack_plan:
+            st.subheader("Task List")
+            for task in attack_plan["task_list"]:
+                st.markdown(f"**Task {task.get('id', '')}:** {task.get('name', '')}")
+                st.markdown(f"*Description:* {task.get('description', '')}")
+                if "commands" in task:
+                    st.code("\n".join(task["commands"]), language="bash")
+                st.markdown("---")
+
+def display_exploit_paths(exploit_paths):
+    """Display the exploit paths in a structured format."""
+    with st.expander("Metasploit Exploit Paths", expanded=True):
+        for exploit in exploit_paths:
+            if "exploit_info" in exploit and "exploit_path" in exploit["exploit_info"]:
+                st.markdown(f"**Exploit for {exploit.get('vulnerability_name', '')}:**")
+                st.code(exploit["exploit_info"]["exploit_path"], language="bash")
+                if "exploit_description" in exploit["exploit_info"]:
+                    st.markdown(f"*Description:* {exploit['exploit_info']['exploit_description']}")
+                st.markdown("---")
