+++ python/ui/pentest_tab.py
@@ -1,0 +1,233 @@
+"""
+Pentest tab UI components for the Streamlit application.
+"""
+
+import streamlit as st
+import time
+import math
+import sys
+import os
+
+# Add the parent directory to sys.path to allow absolute imports
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+from config.settings import CYBERSECURITY_TIPS, SCAN_TYPE_INFO
+
+def render_pentest_tab():
+    """Render the Pentest tab UI components."""
+    st.title("Cybersecurity Orchestrator")
+    
+    # Debug output to verify session state
+    with st.expander("Debug Session State", expanded=False):
+        st.write("Session state scan_progress:", st.session_state.get('scan_progress', 'Not initialized'))
+        if 'scan_progress' in st.session_state:
+            st.write("Running:", st.session_state.scan_progress.get('running', False))
+            st.write("Target:", st.session_state.scan_progress.get('target', ''))
+            st.write("Scan Type:", st.session_state.scan_progress.get('scan_type', ''))
+            st.write("Percent Complete:", st.session_state.scan_progress.get('percent_complete', 0))
+    
+    # Check if a scan is currently running
+    if 'scan_progress' in st.session_state and st.session_state.scan_progress.get('running', False):
+        render_scan_progress()
+    else:
+        st.info("No scan is currently running. Use the chat interface to start a scan.")
+
+def render_scan_progress():
+    """Render the scan progress UI when a scan is running."""
+    # Create a container for the scan progress UI
+    scan_progress_container = st.container()
+    
+    with scan_progress_container:
+        st.subheader(f"Scanning {st.session_state.scan_progress.get('target', 'unknown')} ({st.session_state.scan_progress.get('scan_type', 'unknown')} scan)")
+        
+        # Progress bar
+        st.progress(st.session_state.scan_progress.get('percent_complete', 0) / 100)
+        
+        # Status information
+        col1, col2, col3 = st.columns(3)
+        with col1:
+            st.metric("Current Phase", st.session_state.scan_progress.get('current_phase', 'Unknown'))
+        with col2:
+            st.metric("Open Ports", st.session_state.scan_progress.get('ports_found', 0))
+        with col3:
+            st.metric("Vulnerabilities", st.session_state.scan_progress.get('vulnerabilities_found', 0))
+        
+        # Time information
+        if st.session_state.scan_progress.get('start_time'):
+            elapsed = time.time() - st.session_state.scan_progress.get('start_time')
+            elapsed_str = f"{int(elapsed // 60)}m {int(elapsed % 60)}s"
+            st.info(f"⏱️ Elapsed: {elapsed_str} | ⏳ {st.session_state.scan_progress.get('estimated_time_remaining', 'calculating...')}")
+        
+        # Latest update
+        if st.session_state.scan_progress.get('last_update'):
+            st.success(f"Latest: {st.session_state.scan_progress.get('last_update')}")
+        
+        # Scan log
+        with st.expander("Scan Log", expanded=False):
+            for log_entry in st.session_state.scan_progress.get('log_messages', []):
+                st.text(log_entry)
+        
+        # Scan type information
+        scan_type = st.session_state.scan_progress.get('scan_type', '')
+        if scan_type in SCAN_TYPE_INFO:
+            with st.expander("About This Scan Type", expanded=False):
+                info = SCAN_TYPE_INFO[scan_type]
+                st.markdown(f"### {scan_type.capitalize()} Scan")
+                st.markdown(f"**Description:** {info['description']}")
+                st.markdown(f"**Details:** {info['details']}")
+                st.markdown(f"**Typical Duration:** {info['duration']}")
+                st.markdown(f"**Detection Risk:** {info['detection_risk']}")
+        
+        # Cybersecurity tips
+        with st.expander("Cybersecurity Tips While You Wait", expanded=True):
+            # Display a random tip that changes every few seconds
+            tip_index = int((time.time() // 10) % len(CYBERSECURITY_TIPS))
+            st.info(f"💡 **Tip:** {CYBERSECURITY_TIPS[tip_index]}")
+        
+        # Network visualization
+        with st.expander("Network Visualization", expanded=True):
+            render_network_visualization()
+        
+        # Auto-refresh the UI using a more reliable approach
+        if st.session_state.scan_progress.get('running', False):
+            st.rerun()
+
+def render_network_visualization():
+    """Render the network visualization component."""
+    # Create a simple network visualization using HTML/CSS
+    # Safely get values from session state with defaults
+    ports = st.session_state.scan_progress.get('ports_found', 0)
+    vulnerabilities = st.session_state.scan_progress.get('vulnerabilities_found', 0)
+    target = st.session_state.scan_progress.get('target', 'Target')
+    
+    # Debug output for visualization
+    st.write(f"Rendering visualization with {ports} ports and {vulnerabilities} vulnerabilities")
+    
+    # Generate HTML for the visualization
+    html = f"""
+    <style>
+        .network-viz {{
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            height: 300px;
+            position: relative;
+        }}
+        .target {{
+            width: 100px;
+            height: 100px;
+            background-color: #1E88E5;
+            border-radius: 50%;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            color: white;
+            font-weight: bold;
+            z-index: 10;
+            box-shadow: 0 0 20px rgba(30, 136, 229, 0.5);
+        }}
+        .port {{
+            position: absolute;
+            width: 30px;
+            height: 30px;
+            background-color: #4CAF50;
+            border-radius: 50%;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            color: white;
+            font-size: 10px;
+            box-shadow: 0 0 10px rgba(76, 175, 80, 0.5);
+        }}
+        .vuln {{
+            position: absolute;
+            width: 40px;
+            height: 40px;
+            background-color: #F44336;
+            border-radius: 50%;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            color: white;
+            font-size: 12px;
+            box-shadow: 0 0 15px rgba(244, 67, 54, 0.7);
+            animation: pulse 2s infinite;
+        }}
+        @keyframes pulse {{
+            0% {{ transform: scale(1); }}
+            50% {{ transform: scale(1.1); }}
+            100% {{ transform: scale(1); }}
+        }}
+        .line {{
+            position: absolute;
+            background-color: #90A4AE;
+            height: 2px;
+            transform-origin: 0 0;
+            z-index: 1;
+        }}
+    </style>
+    <div class="network-viz">
+        <div class="target">{target}</div>
+    """
+    
+    # Add ports (green circles)
+    import random
+    random.seed(42)  # Use fixed seed for consistent layout
+    for i in range(min(ports, 10)):  # Show up to 10 ports
+        angle = i * (360 / min(ports, 10)) * (3.14159 / 180)  # Convert to radians
+        distance = 120  # Distance from center
+        x = 50 + distance * math.cos(angle)
+        y = 50 + distance * math.sin(angle)
+        
+        # Calculate line properties
+        length = math.sqrt((x - 50) ** 2 + (y - 50) ** 2)
+        angle_deg = math.atan2(y - 50, x - 50) * (180 / 3.14159)
+        
+        # Add line
+        html += f"""
+        <div class="line" style="
+            width: {length}px;
+            left: 50%;
+            top: 50%;
+            transform: rotate({angle_deg}deg);
+        "></div>
+        """
+        
+        # Add port
+        html += f"""
+        <div class="port" style="left: calc(50% + {x - 15}px); top: calc(50% + {y - 15}px);">
+            {i+1}
+        </div>
+        """
+    
+    # Add vulnerabilities (red circles)
+    for i in range(min(vulnerabilities, 5)):  # Show up to 5 vulnerabilities
+        angle = i * (360 / min(vulnerabilities, 5)) * (3.14159 / 180)  # Convert to radians
+        distance = 180  # Distance from center
+        x = 50 + distance * math.cos(angle)
+        y = 50 + distance * math.sin(angle)
+        
+        # Calculate line properties
+        length = math.sqrt((x - 50) ** 2 + (y - 50) ** 2)
+        angle_deg = math.atan2(y - 50, x - 50) * (180 / 3.14159)
+        
+        # Add line
+        html += f"""
+        <div class="line" style="
+            width: {length}px;
+            left: 50%;
+            top: 50%;
+            transform: rotate({angle_deg}deg);
+        "></div>
+        """
+        
+        # Add vulnerability
+        html += f"""
+        <div class="vuln" style="left: calc(50% + {x - 20}px); top: calc(50% + {y - 20}px);">
+            !
+        </div>
+        """
+    
+    html += "</div>"
+    
+    # Display the visualization
+    st.components.v1.html(html, height=300)
