+++ python/ui/sidebar.py
@@ -1,0 +1,35 @@
+"""
+Sidebar UI components for the Streamlit application.
+"""
+
+import streamlit as st
+import sys
+import os
+
+# Add the parent directory to sys.path to allow absolute imports
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+from modules import tools
+
+def render_sidebar():
+    """Render the sidebar UI components and return the selected model."""
+    st.sidebar.title("Settings")
+    model_option = st.sidebar.selectbox(
+        "Select AI Model",
+        ["Claude (Anthropic)", "GPT-4o mini (OpenAI)"],
+        index=0,
+        key="model_selector"
+    )
+
+    # Add nmap settings in sidebar
+    st.sidebar.divider()
+    st.sidebar.subheader("Nmap Settings")
+    use_docker = st.sidebar.checkbox("Use Docker for Nmap", value=False,
+                                   help="Run nmap in a Docker container instead of locally")
+    st.sidebar.info("All scans use non-privileged techniques that don't require root access")
+
+    tools.USE_DOCKER_NMAP = use_docker
+    
+    # Show which model is active
+    st.sidebar.success(f"Active model: {model_option}")
+    
+    return model_option
