+++ python/ui/testing_tab.py
@@ -1,0 +1,471 @@
+"""
+Testing tab UI components for the Streamlit application.
+"""
+
+import streamlit as st
+import time
+import json
+import re
+import sys
+import os
+from pathlib import Path
+
+# Add the parent directory to sys.path to allow absolute imports
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+def render_testing_tab():
+    """Render the Testing tab UI components."""
+    st.title("Security Analyst Testing")
+    
+    # Import test modules
+    from tests.mock_search import mock_search
+    from modules.security_analyst import analyze_nmap_results, preprocess_json_string
+    
+    # Initialize session state for test results
+    if "test_results" not in st.session_state:
+        st.session_state.test_results = None
+    if "test_running" not in st.session_state:
+        st.session_state.test_running = False
+    
+    # Test configuration
+    st.subheader("Test Configuration")
+    
+    col1, col2 = st.columns(2)
+    
+    with col1:
+        test_type = st.selectbox(
+            "Select Test Type",
+            ["Direct Test", "Mock Search Test", "Real Search Test", "Compare Tests", "Message Wrapper Test", "Real IP Test"],
+            index=0,
+            key="test_type"
+        )
+        
+        sample_file = st.selectbox(
+            "Select Sample Data",
+            ["sample_nmap_xml_output.txt", "Upload Custom File"],
+            index=0,
+            key="sample_file"
+        )
+        
+        if sample_file == "Upload Custom File":
+            uploaded_file = st.file_uploader("Upload Nmap Output File", type=["txt", "xml", "json"])
+        
+    with col2:
+        st.markdown("### Test Options")
+        use_mock = st.checkbox("Use Mock Search", value=True, key="use_mock")
+        show_logs = st.checkbox("Show Debug Logs", value=False, key="show_logs")
+        compare_performance = st.checkbox("Compare Performance", value=False, key="compare_performance")
+    
+    # Message Wrapper Test Configuration UI (shown conditionally)
+    if test_type == "Message Wrapper Test":
+        render_message_wrapper_test_config(preprocess_json_string)
+    
+    # Run test button
+    if st.button("Run Test", type="primary", key="run_test"):
+        run_test(test_type, sample_file, uploaded_file if 'uploaded_file' in locals() else None, 
+                use_mock, mock_search, analyze_nmap_results)
+    
+    # Display test results
+    if st.session_state.test_results:
+        display_test_results(show_logs)
+
+def render_message_wrapper_test_config(preprocess_json_string):
+    """Render the Message Wrapper Test configuration UI."""
+    st.markdown("### Message Wrapper Test Configuration")
+    st.info("This test checks the system's ability to handle LangChain message wrappers like AIMessageChunk.")
+    
+    wrapper_type = st.selectbox(
+        "Select Wrapper Type",
+        ["AIMessageChunk", "HumanMessage", "SystemMessage", "Custom"],
+        index=0,
+        key="wrapper_type"
+    )
+    
+    # Sample JSON content to wrap
+    default_json = """{
+  "format": "json",
+  "scan_type": "vuln",
+  "target": "192.168.1.100",
+  "text_output": "NMAP SCAN RESULTS (vuln scan of 192.168.1.100):\\n==================================================\\n...\\n| smb-vuln-ms17-010: \\n|   VULNERABLE:\\n|   Remote Code Execution vulnerability in Microsoft SMBv1 servers (ms17-010)\\n|     State: VULNERABLE"
+}"""
+    
+    json_content = st.text_area(
+        "JSON Content to Wrap",
+        value=default_json,
+        height=200,
+        key="json_content"
+    )
+    
+    # Create the wrapped content based on selected wrapper type
+    if wrapper_type == "Custom":
+        custom_wrapper = st.text_input(
+            "Custom Wrapper Format (use {} as placeholder for content)",
+            value="CustomMessage(content={})",
+            key="custom_wrapper"
+        )
+        wrapped_content = custom_wrapper.format(f'"{json_content}"')
+    else:
+        wrapped_content = f'{wrapper_type}(content="{json_content}")'
+    
+    # Show the wrapped content
+    with st.expander("View Wrapped Content", expanded=False):
+        st.code(wrapped_content, language="python")
+    
+    # Store the wrapped content in session state for use in run_test
+    st.session_state.wrapped_content = wrapped_content
+
+def run_test(test_type, sample_file, uploaded_file, use_mock, mock_search, analyze_nmap_results):
+    """Run the selected test and store results in session state."""
+    st.session_state.test_running = True
+    
+    with st.spinner("Running test..."):
+        try:
+            # For non-Message Wrapper tests, load sample data
+            if test_type != "Message Wrapper Test":
+                if sample_file == "Upload Custom File" and uploaded_file is not None:
+                    sample_data = uploaded_file.read().decode()
+                else:
+                    sample_file_path = Path(__file__).resolve().parent.parent / "tests" / sample_file
+                    with open(sample_file_path, "r") as f:
+                        sample_data = f.read()
+            
+            # Process based on test type
+            if test_type == "Message Wrapper Test":
+                run_message_wrapper_test()
+            
+            elif test_type == "Real IP Test":
+                run_real_ip_test(use_mock, mock_search, analyze_nmap_results)
+            
+            elif test_type == "Direct Test":
+                start_time = time.time()
+                if use_mock:
+                    result = analyze_nmap_results(sample_data, custom_search_func=mock_search)
+                else:
+                    result = analyze_nmap_results(sample_data)
+                duration = time.time() - start_time
+                
+                # Store results in session state
+                st.session_state.test_results = {
+                    "result": result,
+                    "duration": duration,
+                    "test_type": test_type,
+                    "use_mock": use_mock,
+                    "sample_file": sample_file
+                }
+            
+            elif test_type == "Compare Tests":
+                run_comparison_test(sample_data, mock_search, analyze_nmap_results)
+            
+            else:
+                # Default to direct test
+                start_time = time.time()
+                if use_mock:
+                    result = analyze_nmap_results(sample_data, custom_search_func=mock_search)
+                else:
+                    result = analyze_nmap_results(sample_data)
+                duration = time.time() - start_time
+                
+                # Store results in session state
+                st.session_state.test_results = {
+                    "result": result,
+                    "duration": duration,
+                    "test_type": test_type,
+                    "use_mock": use_mock,
+                    "sample_file": sample_file
+                }
+            
+        except Exception as e:
+            st.error(f"Error running test: {str(e)}")
+            st.session_state.test_results = {
+                "error": str(e),
+                "test_type": test_type,
+                "use_mock": use_mock,
+                "sample_file": sample_file
+            }
+        
+        st.session_state.test_running = False
+
+def run_message_wrapper_test():
+    """Run the Message Wrapper Test."""
+    from modules.security_analyst import preprocess_json_string
+    
+    start_time = time.time()
+    wrapped_content = st.session_state.wrapped_content
+    preprocessed = preprocess_json_string(wrapped_content)
+    
+    # Try to parse the preprocessed JSON
+    try:
+        parsed_json = json.loads(preprocessed)
+        
+        # Store results
+        result = {
+            "success": True,
+            "preprocessing_time": time.time() - start_time,
+            "parsed_json": parsed_json,
+            "wrapper_type": st.session_state.wrapper_type,
+            "original_wrapped": wrapped_content,
+            "preprocessed": preprocessed,
+            "is_message_wrapper_test": True
+        }
+    except json.JSONDecodeError as e:
+        result = {
+            "success": False,
+            "error": f"JSON parsing error: {str(e)}",
+            "error_position": e.pos,
+            "error_context": preprocessed[max(0, e.pos-50):min(len(preprocessed), e.pos+50)],
+            "wrapper_type": st.session_state.wrapper_type,
+            "original_wrapped": wrapped_content,
+            "preprocessed": preprocessed,
+            "is_message_wrapper_test": True
+        }
+    
+    # Store results in session state
+    st.session_state.test_results = {
+        "result": result,
+        "duration": time.time() - start_time,
+        "test_type": "Message Wrapper Test"
+    }
+
+def run_real_ip_test(use_mock, mock_search, analyze_nmap_results):
+    """Run the Real IP Test."""
+    # Load the sample data for 10.0.255.165 from the test_real_ip.py script
+    real_ip_sample_path = Path(__file__).resolve().parent.parent / "test_real_ip.py"
+    
+    # Extract the sample data from the test_real_ip.py file
+    with open(real_ip_sample_path, "r") as f:
+        file_content = f.read()
+        # Extract the sample data between triple quotes
+        sample_data_match = re.search(r'sample_data = """(.*?)"""', file_content, re.DOTALL)
+        if sample_data_match:
+            sample_data = sample_data_match.group(1)
+            st.info("Testing with real IP: 10.0.255.165")
+        else:
+            st.error("Could not extract sample data from test_real_ip.py")
+            sample_data = "Error: Could not extract sample data"
+    
+    # Set start time for duration calculation
+    start_time = time.time()
+    
+    # Analyze the sample data
+    if use_mock:
+        result = analyze_nmap_results(sample_data, custom_search_func=mock_search)
+    else:
+        result = analyze_nmap_results(sample_data)
+    
+    # Store results in session state
+    st.session_state.test_results = {
+        "result": result,
+        "duration": time.time() - start_time,
+        "test_type": "Real IP Test",
+        "use_mock": use_mock
+    }
+
+def run_comparison_test(sample_data, mock_search, analyze_nmap_results):
+    """Run both real and mock tests and compare."""
+    # Run both real and mock tests and compare
+    mock_start_time = time.time()
+    mock_result = analyze_nmap_results(sample_data, custom_search_func=mock_search)
+    mock_duration = time.time() - mock_start_time
+    
+    real_start_time = time.time()
+    real_result = analyze_nmap_results(sample_data)
+    real_duration = time.time() - real_start_time
+    
+    result = {
+        "mock_result": mock_result,
+        "real_result": real_result,
+        "mock_duration": mock_duration,
+        "real_duration": real_duration,
+        "is_comparison": True
+    }
+    
+    # Store results in session state
+    st.session_state.test_results = {
+        "result": result,
+        "duration": mock_duration + real_duration,
+        "test_type": "Compare Tests"
+    }
+
+def display_test_results(show_logs):
+    """Display the test results."""
+    st.subheader("Test Results")
+    
+    results = st.session_state.test_results
+    
+    if "error" in results:
+        st.error(f"Test failed: {results['error']}")
+    else:
+        # Display test metadata
+        st.info(f"Test completed in {results['duration']:.2f} seconds")
+        
+        if "is_message_wrapper_test" in results["result"] and results["result"]["is_message_wrapper_test"]:
+            display_message_wrapper_test_results(results["result"])
+        elif "is_comparison" in results["result"] and results["result"]["is_comparison"]:
+            display_comparison_test_results(results["result"])
+        else:
+            display_regular_test_results(results["result"], show_logs)
+
+def display_message_wrapper_test_results(wrapper_result):
+    """Display the Message Wrapper Test results."""
+    st.markdown("### Message Wrapper Test Results")
+    
+    if wrapper_result.get("success", False):
+        st.success(f"✅ Successfully processed {wrapper_result['wrapper_type']} wrapper in {wrapper_result['preprocessing_time']:.4f} seconds")
+        
+        # Display the parsed JSON
+        st.subheader("Parsed JSON")
+        st.json(wrapper_result["parsed_json"])
+        
+        # Show before and after
+        col1, col2 = st.columns(2)
+        
+        with col1:
+            st.markdown("#### Original Wrapped Content")
+            st.code(wrapper_result["original_wrapped"], language="python")
+        
+        with col2:
+            st.markdown("#### Preprocessed Content")
+            st.code(wrapper_result["preprocessed"], language="json")
+        
+        # Verify key fields were preserved
+        if "parsed_json" in wrapper_result:
+            st.subheader("Field Verification")
+            parsed = wrapper_result["parsed_json"]
+            
+            # Check for common fields
+            fields_to_check = ["format", "target", "scan_type"]
+            for field in fields_to_check:
+                if field in parsed:
+                    st.success(f"✅ Field '{field}' was preserved: {parsed[field]}")
+                else:
+                    st.error(f"❌ Field '{field}' was not preserved")
+    else:
+        st.error(f"❌ Failed to process {wrapper_result['wrapper_type']} wrapper")
+        st.error(wrapper_result.get("error", "Unknown error"))
+        
+        if "error_context" in wrapper_result:
+            st.markdown("#### Error Context")
+            st.code(f"Error at position {wrapper_result['error_position']}:\n{wrapper_result['error_context']}")
+        
+        if "preprocessed" in wrapper_result:
+            st.markdown("#### Preprocessed Content (Invalid JSON)")
+            st.code(wrapper_result["preprocessed"])
+        
+        # Show the original wrapped content
+        st.markdown("#### Original Wrapped Content")
+        st.code(wrapper_result["original_wrapped"], language="python")
+
+def display_comparison_test_results(comparison):
+    """Display the Comparison Test results."""
+    st.markdown("### Performance Comparison")
+    col1, col2 = st.columns(2)
+    
+    with col1:
+        st.markdown("#### Mock Search")
+        st.info(f"Duration: {comparison['mock_duration']:.2f} seconds")
+    
+    with col2:
+        st.markdown("#### Real Search")
+        st.info(f"Duration: {comparison['real_duration']:.2f} seconds")
+    
+    st.markdown("### MITRE ATT&CK Mappings Comparison")
+    
+    col1, col2 = st.columns(2)
+    
+    with col1:
+        st.markdown("#### Mock Search Mappings")
+        if "mitre_mappings" in comparison["mock_result"]:
+            for i, mapping in enumerate(comparison["mock_result"]["mitre_mappings"]):
+                st.markdown(f"**{i+1}. Technique:** {mapping.get('technique_id')} - {mapping.get('technique_name')}")
+                st.markdown(f"**Tactic:** {mapping.get('tactic')}")
+                st.markdown("---")
+    
+    with col2:
+        st.markdown("#### Real Search Mappings")
+        if "mitre_mappings" in comparison["real_result"]:
+            for i, mapping in enumerate(comparison["real_result"]["mitre_mappings"]):
+                st.markdown(f"**{i+1}. Technique:** {mapping.get('technique_id')} - {mapping.get('technique_name')}")
+                st.markdown(f"**Tactic:** {mapping.get('tactic')}")
+                st.markdown("---")
+    
+    # Display vulnerabilities
+    st.markdown("### Vulnerabilities")
+    if "vulnerabilities" in comparison["mock_result"]:
+        with st.expander("View Vulnerabilities", expanded=True):
+            st.json(comparison["mock_result"]["vulnerabilities"])
+    
+    # Display attack plan
+    st.markdown("### Attack Plan")
+    if "attack_plan" in comparison["mock_result"]:
+        with st.expander("View Attack Plan", expanded=True):
+            # Display executive summary
+            if "executive_summary" in comparison["mock_result"]["attack_plan"]:
+                st.subheader("Executive Summary")
+                st.write(comparison["mock_result"]["attack_plan"]["executive_summary"])
+            
+            # Display task list
+            if "task_list" in comparison["mock_result"]["attack_plan"]:
+                st.subheader("Task List")
+                for task in comparison["mock_result"]["attack_plan"]["task_list"]:
+                    st.markdown(f"**Task {task.get('id', '')}:** {task.get('name', '')}")
+                    st.markdown(f"*Description:* {task.get('description', '')}")
+                    if "commands" in task:
+                        st.code("\n".join(task["commands"]), language="bash")
+                    st.markdown("---")
+
+def display_regular_test_results(result, show_logs):
+    """Display regular test results."""
+    # Display vulnerabilities
+    if "vulnerabilities" in result:
+        st.markdown("### Vulnerabilities Found")
+        with st.expander("View Vulnerabilities", expanded=True):
+            for i, vuln in enumerate(result["vulnerabilities"]):
+                st.markdown(f"**{i+1}. {vuln.get('name')}**")
+                st.markdown(f"*Description:* {vuln.get('description')}")
+                st.markdown(f"*ID:* {vuln.get('id')}")
+                st.markdown("---")
+    
+    # Display MITRE mappings
+    if "mitre_mappings" in result:
+        st.markdown("### MITRE ATT&CK Mappings")
+        with st.expander("View MITRE Mappings", expanded=True):
+            for i, mapping in enumerate(result["mitre_mappings"]):
+                st.markdown(f"**{i+1}. Technique:** {mapping.get('technique_id')} - {mapping.get('technique_name')}")
+                st.markdown(f"**Tactic:** {mapping.get('tactic')}")
+                st.markdown("---")
+    
+    # Display attack plan
+    if "attack_plan" in result:
+        st.markdown("### Attack Plan")
+        with st.expander("View Attack Plan", expanded=True):
+            # Display executive summary
+            if "executive_summary" in result["attack_plan"]:
+                st.subheader("Executive Summary")
+                st.write(result["attack_plan"]["executive_summary"])
+            
+            # Display detailed plan
+            if "detailed_plan" in result["attack_plan"]:
+                st.subheader("Detailed Plan")
+                st.markdown(result["attack_plan"]["detailed_plan"])
+            
+            # Display task list
+            if "task_list" in result["attack_plan"]:
+                st.subheader("Task List")
+                for task in result["attack_plan"]["task_list"]:
+                    st.markdown(f"**Task {task.get('id', '')}:** {task.get('name', '')}")
+                    st.markdown(f"*Description:* {task.get('description', '')}")
+                    if "commands" in task:
+                        st.code("\n".join(task["commands"]), language="bash")
+                    st.markdown("---")
+    
+    # Display raw scan output
+    if "raw_scan_output" in result:
+        st.markdown("### Raw Scan Output")
+        with st.expander("View Raw Scan Output", expanded=False):
+            st.code(result["raw_scan_output"], language="bash")
+    
+    # Display raw result for debugging
+    if show_logs:
+        st.markdown("### Debug Information")
+        with st.expander("View Raw Result Object", expanded=False):
+            st.json(result)
